% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clam.R
\name{accrate.depth}
\alias{accrate.depth}
\title{Title #TBW}
\usage{
accrate.depth(depth, yrcm = TRUE, prob = 0.95)
}
\arguments{
\item{depth}{#TBW}

\item{yrcm}{#TBW}

\item{prob}{#TBW}
}
\value{
#TBW
}
\description{
Title #TBW
}
\examples{
 #TBW
}

