% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clam.R
\name{age.pMC}
\alias{age.pMC}
\title{Title #TBW}
\usage{
age.pMC(mn, sdev, ratio = 100, decimals = 3)
}
\arguments{
\item{mn}{#TBW}

\item{sdev}{#TBW}

\item{ratio}{#TBW}

\item{decimals}{#TBW}
}
\value{
#TBW
}
\description{
Title #TBW
}
\examples{
 #TBW
}

