% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr.R
\name{agebasis2clam}
\alias{agebasis2clam}
\title{Reshape no-C14 data to CLAM format}
\usage{
agebasis2clam(agebasis)
}
\arguments{
\item{noC14}{Data frame with no-C14 data as those in a chronology list returned by \code{\link[EPDr:getChronologies]{getChronologies}}.}
}
\value{
Data frame with no-C14 data in CLAM format. This data frame can be easily combined with C14 data from \code{\link[EPDr:getChronologies]{c142clam}} using \code{rbind}.
}
\description{
This function takes no-C14 data, that can be extracted from a chronology object, to fit into a new table that comply with
CLAM format.
}
\examples{
library(EPDr)
epd.connection <- connectToEPD(host="diegonl.ugr.es", database="epd_ddbb",
                               user="epdr", password="epdrpw")
c14 <- getC14(400, epd.connection)
c14.clam <- c142clam(c14)
chron <- getChronologies(400, epd.connection)
noc14.clam <- agebasis2clam(chron$agebasis)
all.clam <- rbind(c14.clam, noc14.clam)
}

