% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr-export_functions.R
\name{agebasis_export}
\alias{agebasis_export}
\title{Title Reshape no-C14 data to CLAM or BACON format}
\usage{
agebasis_export(agebasis, format = c("clam", "bacon"))
}
\arguments{
\item{agebasis}{Data frame with no-C14 data as those in a chronology list returned by
\code{\link[EPDr:getChronology]{getChronology}}.}

\item{format}{Character string indicating whether to export to "clam" or to "bacon" format.}
}
\value{
Data frame with no-C14 data in CLAM or BACON format. This data frame can be easily combined with C14
data from \code{\link[EPDr:c14_export]{c14_export}} using \code{rbind}.
}
\description{
This function takes no-C14 data, that can be extracted from a chronology object, to fit into a new table that
comply with CLAM or BACON format.
}
\examples{
#library(EPDr)
#epd.connection <- connectToEPD(host="localhost", database="epd_ddbb",
#                               user="epdr", password="epdrpw")
#c14 <- getC14(400, epd.connection)
#c14.clam <- c14_export(c14, "clam")
#c14.bacon <- c14_export(c14, "bacon")
#chron <- getChronology(400, epd.connection)
#noc14.clam <- agebasis_export(chron@agebasis, "clam")
#noc14.bacon <- agebasis_export(chron@agebasis, "bacon")
#all.clam <- rbind(c14.clam, noc14.clam)
#all.bacon <- rbind(c14.bacon, noc14.bacon)
}

