% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr-classes.R
\docType{class}
\name{agedcounts-class}
\alias{agedcounts}
\alias{agedcounts-class}
\title{Title TBW}
\value{
TBW
}
\description{
Title TBW
}
\section{Slots}{

\describe{
\item{\code{e_}}{numeric. TBW}

\item{\code{restriction}}{data.frame. TBW}

\item{\code{entity}}{TBW}

\item{\code{site}}{data.frame. TBW}

\item{\code{ages}}{ages. TBW}

\item{\code{counts}}{counts. TBW}
}}
\examples{
# TBW
}

