% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr-classes.R
\docType{class}
\name{ages-class}
\alias{ages}
\alias{ages-class}
\title{Title TBW}
\value{
TBW
}
\description{
Title TBW
}
\section{Slots}{

\describe{
\item{\code{e_}}{numeric. TBW}

\item{\code{restriction}}{data.frame. TBW}

\item{\code{entity}}{TBW}

\item{\code{site}}{data.frame. TBW}

\item{\code{default_chronology}}{numeric. TBW}

\item{\code{giesecke}}{logical. TBW}

\item{\code{sample_}}{numeric. TBW}

\item{\code{sample_label}}{character. TBW}

\item{\code{depthcm}}{numeric. TBW}

\item{\code{depths}}{data.frame. TBW}

\item{\code{depth_ages}}{data.frame. TBW}

\item{\code{data_quality}}{data.frame. TBW}
}}
\examples{
# TBW
}

