% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr-classes.R
\docType{class}
\name{agesdf-class}
\alias{agesdf-class}
\alias{agesdf}
\title{Class for agesdf objects}
\description{
This object stores information regarding the ages of biological 
samples in an entity but in vector or table format rather than 
the original format of the EPD database. Note, however, that data 
here comes from the tables \code{pagedpt} and \code{psamples} from
the database or an \code{\link[EPDr]{epd.entity}} object. 
Contrary to data in \code{\link[EPDr]{epd.entity}} objects, 
this are intended to be manipulated and modified if necessary.
}
\section{Slots}{

\describe{
\item{\code{depthcm}}{numeric. Depth of the samples in cm.}

\item{\code{depthages}}{data.frame. Estimated ages for each sample (rows) 
according to each chronology (columns), including chronologies
from Giesecke et al. (2013) if available for that entity.}

\item{\code{dataquality}}{data.frame. This dataframe include a quality index
for each sample. By default it is empty, but it can be calculated
by means of specific functions like \code{\link[EPDr]{blois_quality}}.}
}}

