% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr-standardize_methods.R
\docType{methods}
\name{blois_quality}
\alias{blois_quality}
\alias{blois_quality,epd.entity.df-method}
\alias{blois_quality,epd.entity-method}
\title{Blois quality index for palynological samples}
\usage{
blois_quality(x, chron = NULL, max_sample_dist = 2000,
  max_control_dist = 5000)

\S4method{blois_quality}{epd.entity.df}(x, chron = NULL,
  max_sample_dist = 2000, max_control_dist = 5000)

\S4method{blois_quality}{epd.entity}(x, chron = NULL,
  max_sample_dist = 2000, max_control_dist = 5000)
}
\arguments{
\item{x}{epd.entity \code{\link[EPDr]{epd.entity.df}} or a
\code{\link[EPDr]{epd.entity}} object.}

\item{chron}{numeric Chronology number to look for samples and control points ages.
This value become the default chronology in the new object. If not 
specified the function check the default chronology in  
\code{x}. It can be any of the chronologies in the EPD for that 
particular entity, or the one from Giesecke et al. (2013).}

\item{max_sample_dist}{numeric Maximum numeric distance in years to be considered
to the palynological samples for interpolated or ranged data.}

\item{max_control_dist}{numeric Maximum numeric distance in years to be considered
to the control points (e.g., C14, top, bottom, etc.).}
}
\value{
\code{\link[EPDr]{epd.entity.df}} object with no empty
\code{@agesdf@dataquality} slot. The default chronology in \code{x@defaultchron}
is changed to the one specified in \code{chron}.
}
\description{
This function apply the quality index described in Blois et al. (2013). 
From the Ecography 2013 paper, Appendix 3: "\emph{For each site at a particular 1 kyr
time period, site data-quality was calculated as the mean normalized distance
of the nearest pollen sample and the nearest chronological control.  We
calculated the distance in years of the nearest pollen sample and the nearest
chronological control to each 1 kyr time period.  We eliminated sites where
the nearest pollen sample was over 2000 years away or the nearest
chronological control was over 5000 years away.  For the remaining sites in
each 1 kyr period, we created a summary measure of site data-quality by 
rescaling the two distances in years to a 0 - 1 scale and calculating the
mean.  For example, if the nearest sample to the 1 kyr BP time period at a
given site was at 1.050 kyr BP and the nearest chronological control was at
1.100 kyr BP, the raw distances would be 50 years and 100 years, respectively.
These equate to scaled values of 0.975 (i.e., 1 - 50/2000) and 0.98 (i.e., 1 -
100/5000) for sample and chronological quality, respectively, with a mean
data-quality for this site at the 1 kyr BP time period of 0.9775.}"
To replicate the calculation the function allows to specify different maximum
distances as parameters of the function.
}
\details{
When \code{x} is an \code{\link[EPDr]{epd.entity}} object, it is
first transformed into a \code{\link[EPDr]{epd.entity.df}} 
object by \code{\link[EPDr]{entity_to_matrices}} function.
}
\examples{
\dontrun{
epd.connection <- connect_to_epd(host="localhost", database="epd",
                                 user="epdr", password="epdrpw")
epd.1 <- get_entity(1, epd.connection)
epd.1.qi <- blois_quality(epd.1)
epd.1.qi@agesdf@dataquality

epd.1.ran <- intervals_counts(epd.1, tmin = seq(0, 21000, by = 1000),
                              tmax = seq(999, 21999, by = 1000))
epd.1.ran.qi <- blois_quality(epd.1.ran)
epd.1.ran.qi@agesdf@dataquality

t <- c(seq(0, 21000, by = 500))
epd.1.int <- interpolate_counts(epd.1, t)
epd.1.int.qi <- blois_quality(epd.1.int)
epd.1.int.qi@agesdf@dataquality
}
}
\references{
Blois, J.L., Williams, J.W., Fitzpatrick, M.C., Ferrier, S.,
Veloz, S.D., He, F., Liu, Z., Manion, G., and Otto-Bliesner, B. (2013).
Modeling the climatic drivers of spatial patterns in vegetation composition
since the Last Glacial Maximum. Ecography, 36, 460-473.

Giesecke, T., Davis, B., Brewer, S., Finsinger, W., 
Wolters, S., Blaaw, M., de Beaulieu, J.L., Binney, H., Fyfe, R.M.,
Gaillard, M.J., Gil-Romera, G., van der Knaap, W.O. Kunes, P.,
Kuhl, N., van Leeuwen, J.F.N, Leydet, M., Lotter, A.F., Ortu, E.,
Semmler, M., and Bradshaw, R.H.W (2013). Towards mapping the late
Quaternary vegetation change of Europe. Vegetation History and
Archaeobotany, 23, 75-86.
}
