% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr-standardize_functions.R
\name{bloisQualityIndex}
\alias{bloisQualityIndex}
\title{Blois quality index for palynological samples}
\usage{
bloisQualityIndex(agedcounts, datation, max_sample_dist = 2000,
  max_c14_dist = 5000)
}
\arguments{
\item{agedcounts}{\code{\link[EPDr:agedcounts]{agedcounts}} object.}

\item{datation}{\code{\link[EPDr:datation]{datation}} object for the same entity
in \code{agedcounts}. This is mandatory to calculate distances between samples and 
control points.}

\item{max_sample_dist}{Maximum numeric distance to be considered to the
palynological samples for interpolated or ranged data.}

\item{max_c14_dist}{Maximum numeric distance to be considered to the (C14) control
points.}
}
\value{
\code{\link[EPDr:agedcounts]{agedcounts}} object with no empty
\code{@ages@data_quality} slot.
}
\description{
This function apply the quality index described in Blois et al. (2013).
}
\details{
From the Ecography 2013 paper, Appendix 3: "For each site at a particular 1 kyr
time period, site data-quality was calculated as the mean normalized distance
of the nearest pollen sample and the nearest chronological control.  We
calculated the distance in years of the nearest pollen sample and the nearest
chronological control to each 1 kyr time period.  We eliminated sites where
the nearest pollen sample was over 2000 years away or the nearest
chronological control was over 5000 years away.  For the remaining sites in
each 1 kyr period, we created a summary measure of site data-quality by 
rescaling the two distances in years to a 0 - 1 scale and calculating the
mean.  For example, if the nearest sample to the 1 kyr BP time period at a
given site was at 1.050 kyr BP and the nearest chronological control was at
1.100 kyr BP, the raw distances would be 50 years and 100 years, respectively.
These equate to scaled values of 0.975 (i.e., 1 - 50/2000) and 0.98 (i.e., 1 -
100/5000) for sample and chronological quality, respectively, with a mean
data-quality for this site at the 1 kyr BP time period of 0.9775."

To replicate the calculation the function allows to specify different maximum
distances as parameters of the function.
}
\examples{
# epd.connection <- connectToEPD(host="localhost", database="epd",
# user="epdr", password="epdrpw")
# agedcounts.1 <- getAgedCounts(1, epd.connection)
# datation.1 <- getDatation(1, epd.connection)
# agedcounts.1.qi <- bloisQualityIndex(agedcounts.1, datation.1)
# agedcounts.1.qi@ages@data_quality
# 
# agedcounts.1.ran <- intervalsCounts(agedcounts.1, tmin=seq(0, 21000, by=1000),
# tmax=seq(999, 21999, by=1000))
# agedcounts.1.ran.qi <- bloisQualityIndex(agedcounts.1.ran, datation.1)
# agedcounts.1.ran.qi@ages@data_quality
# 
# t <- c(seq(0, 21000, by=500))
# agedcounts.1.int <- interpolateCounts(agedcounts.1, t)
# agedcounts.1.int.qi <- bloisQualityIndex(agedcounts.1.int, datation.1)
# agedcounts.1.int.qi@ages@data_quality
}
\references{
Blois, J.L, J.W. Williams, M.C. Fitzpatrick, S. Ferrier, S.D. Veloz, F. He, Z. Liu, G. Manion, and Bette Otto-Bliesner (2013). Modeling the Climatic Drivers of Spatial Patterns in Vegetation Composition since the Last Glacial Maximum. Ecography 36(4): 460-473. doi:10.1111/j.1600-0587.2012.07852.x.
}

