% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr-epd2clam_functions.R
\name{c14_epd2clam}
\alias{c14_epd2clam}
\title{Reshape C14 data to CLAM format}
\usage{
c14_epd2clam(C14)
}
\arguments{
\item{C14}{Data frame with C14 data as those extracted from \code{\link[EPDr:getC14]{getC14}}.}
}
\value{
Data frame with C14 data in CLAM format.
}
\description{
This function takes C14 data, as those extracted \code{\link[EPDr:getC14]{getC14}}, to fit into a new table that comply with
CLAM format.
}
\examples{
library(EPDr)
epd.connection <- connectToEPD(host="diegonl.ugr.es", database="epd_ddbb",
                               user="epdr", password="epdrpw")
c14 <- getC14(400, epd.connection)
c14.clam <- c14_epd2clam(c14)
}

