% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clam.R
\name{calibrate}
\alias{calibrate}
\title{Title}
\usage{
calibrate(cage = 2450, error = 50, reservoir = 0, prob = 0.95, cc = 1,
  cc1 = "IntCal13.14C", cc2 = "Marine13.14C", cc3 = "SHCal13.14C",
  cc4 = "mixed.14C", cc5 = "gluedHemispheres.14C", postbomb = FALSE,
  pb1 = "postbomb_NH1.14C", pb2 = "postbomb_NH2.14C",
  pb3 = "postbomb_NH3.14C", pb4 = "postbomb_SH1_2.14C",
  pb5 = "postbomb_SH3.14C", yrsteps = 1, pbsteps = 0.01, hpdsteps = 1,
  calibt = FALSE, yrmin = c(), yrmax = c(), minC14 = c(),
  maxC14 = c(), times = 5, calheight = 0.3, expand = 0.1,
  threshold = 1e-06, storedat = FALSE, graph = TRUE, xlab = c(),
  ylab = c(), BCAD = FALSE, mar = c(3.5, 3, 2, 1), mgp = c(2, 1, 0),
  bty = "l", title = c(), date.col = "red", cc.col = rgb(0, 0.5, 0,
  0.7), dist.col = rgb(0, 0, 0, 0.3), sd.col = rgb(0, 0, 0, 0.5))
}
\arguments{
\item{cage}{}

\item{error}{}

\item{reservoir}{}

\item{prob}{}

\item{cc}{}

\item{cc1}{}

\item{cc2}{}

\item{cc3}{}

\item{cc4}{}

\item{cc5}{}

\item{postbomb}{}

\item{pb1}{}

\item{pb2}{}

\item{pb3}{}

\item{pb4}{}

\item{pb5}{}

\item{yrsteps}{}

\item{pbsteps}{}

\item{hpdsteps}{}

\item{calibt}{}

\item{yrmin}{}

\item{yrmax}{}

\item{minC14}{}

\item{maxC14}{}

\item{times}{}

\item{calheight}{}

\item{expand}{}

\item{threshold}{}

\item{storedat}{}

\item{graph}{}

\item{xlab}{}

\item{ylab}{}

\item{BCAD}{}

\item{mar}{}

\item{mgp}{}

\item{bty}{}

\item{title}{}

\item{date.col}{}

\item{cc.col}{}

\item{dist.col}{}

\item{sd.col}{}
}
\value{

}
\description{
Title
}

