% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr-standardize_methods.R
\docType{methods}
\name{check_counts}
\alias{check_counts}
\alias{check_counts,epd.entity.df-method}
\title{Check counts data on EPDr objects}
\usage{
check_counts(x)

\S4method{check_counts}{epd.entity.df}(x)
}
\arguments{
\item{x}{epd.entity.df \code{\link[EPDr]{epd.entity.df}} object.}
}
\value{
Logical value indicating whether the object has a default 
chronology (TRUE) or not (FALSE).
}
\description{
The function check \code{\link[EPDr]{epd.entity.df}} objects (with
slot @commdf@counts) to see if there are counts data on it.
}
\examples{
\dontrun{
epd.connection <- connect_to_epd(host="localhost", database="epd",
                                 user="epdr", password="epdrpw")
epd.1 <- get_entity(1, epd.connection)
check_counts(epd.1)
epd.763 <- get_entity(763, epd.connection)
check_counts(epd.763)
}
}
