% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr-standardize_methods.R
\docType{methods}
\name{check_default_chron}
\alias{check_default_chron}
\alias{check_default_chron,epd.entity-method}
\alias{check_default_chron-method}
\title{Check default chronology on EPDr objects}
\usage{
check_default_chron(x)

\S4method{check_default_chron}{epd.entity}(x)
}
\arguments{
\item{x}{\code{\link[EPDr:epd.entity]{epd.entity}} or
\code{\link[EPDr:epd.entity.df]{epd.entity.df}} objects.}
}
\value{
Logical value indicating whether the object has a default chronology (TRUE)
or not (FALSE).
}
\description{
The function check EPDr objects with slot @defaultchron 
(\code{\link[EPDr:epd.entity]{epd.entity}}, and
\code{\link[EPDr:epd.entity.df]{epd.entity.df}}) to see if there are a default
chronology specified.
}
\details{
Details TBW
}
\examples{
# epd.connection <- connectToEPD(host="localhost", database="epd",
#                                user="epdr", password="epdrpw")
# e.list <- listE(epd.connection, country = c("Spain"))
# e.list <- e.list$e_
# spain.agedcounts <- lapply(e.list, getAgedCounts, epd.connection)
# spain.agedcounts.wo <- sapply(spain.agedcounts, check_default_chron)
# spain.agedcounts.wo
# spain.agedcounts[[1]]@ages@default_chronology
# spain.agedcounts[[16]]@ages@default_chronology
}
