% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr-standardize_methods.R
\docType{methods}
\name{check_default_chron}
\alias{check_default_chron}
\alias{check_default_chron,epd.entity-method}
\title{Check default chronology on EPDr objects}
\usage{
check_default_chron(x)

\S4method{check_default_chron}{epd.entity}(x)
}
\arguments{
\item{x}{epd.entity \code{\link[EPDr]{epd.entity}} or
\code{\link[EPDr]{epd.entity.df}} objects.}
}
\value{
Logical value indicating whether the object has a default 
chronology (TRUE) or not (FALSE).
}
\description{
The function check EPDr objects with slot @defaultchron 
(\code{\link[EPDr]{epd.entity}}, and
\code{\link[EPDr]{epd.entity.df}}) to see if there are a default
chronology specified.
}
\examples{
# epd.connection <- connect_to_epd(host="localhost", database="epd",
#                                user="epdr", password="epdrpw")
# epd.1 <- get_entity(1, epd.connection)
# check_default_chron(epd.1)
}
