% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr-standardize_methods.R
\docType{methods}
\name{check_restriction}
\alias{check_restriction}
\alias{check_restriction,epd.entity-method}
\alias{check_restriction-method}
\title{Check restrictions on EPDr objects}
\usage{
check_restriction(x)

\S4method{check_restriction}{epd.entity}(x)
}
\arguments{
\item{x}{\code{\link[EPDr:epd.entity]{epd.entity}} or
\code{\link[EPDr:epd.entity.df]{epd.entity.df}} object.}
}
\value{
logical value indicating whether the data are restricted (TRUE) or
unrestricted (FALSE).
}
\description{
This function check EPDr objects (\code{\link[EPDr:epd.entity]{epd.entity}},
and \code{\link[EPDr:epd.entity.df]{epd.entity.df}}), 
and return a logical value indicating whether the data are restricted (TRUE) or
unrestricted (FALSE).
}
\details{
Details TBW
}
\examples{
# epd.connection <- connectToEPD(host="localhost", database="epd",
# user="epdr", password="epdrpw")
# agedcount <- getAgedCounts(1, epd.connection)
# check_restriction(agedcount)
# agedcount <- getAgedCounts(1046, epd.connection)
# check_restriction(agedcount)
}
