% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr-standardize_methods.R
\docType{methods}
\name{check_restriction}
\alias{check_restriction}
\alias{check_restriction,epd.entity-method}
\title{Check restrictions on EPDr objects}
\usage{
check_restriction(x)

\S4method{check_restriction}{epd.entity}(x)
}
\arguments{
\item{x}{epd.entity \code{\link[EPDr]{epd.entity}} or
\code{\link[EPDr]{epd.entity.df}} object.}
}
\value{
logical value indicating whether the data are restricted (TRUE) or
unrestricted (FALSE).
}
\description{
This function check EPDr objects (\code{\link[EPDr]{epd.entity}},
and \code{\link[EPDr]{epd.entity.df}}), 
and return a logical value indicating whether the data are restricted (TRUE) or
unrestricted (FALSE).
}
\examples{
\dontrun{
epd.connection <- connect_to_epd(host="localhost", database="epd",
                                 user="epdr", password="epdrpw")
epd.1 <- get_entity(1, epd.connection)
check_restriction(epd.1)
epd.1046 <- get_entity(1046, epd.connection)
check_restriction(epd.1046)
}
}
