% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr-standardize_methods.R
\docType{methods}
\name{check_valid_defaultchron}
\alias{check_valid_defaultchron}
\alias{check_valid_defaultchron,epd.entity-method}
\title{Check that default chronology on EPDr objects is valid}
\usage{
check_valid_defaultchron(x)

\S4method{check_valid_defaultchron}{epd.entity}(x)
}
\arguments{
\item{x}{epd.entity \code{\link[EPDr]{epd.entity-class}} or
\code{\link[EPDr]{epd.entity.df-class}} objects.}
}
\value{
Logical value indicating whether the default chronology has ages 
(TRUE) or not (FALSE).
}
\description{
The function check EPDr objects with slot @defaultchron 
(\code{\link[EPDr]{epd.entity-class}}, and
\code{\link[EPDr]{epd.entity.df-class}}) to see if that chronology
has ages for the counts samples.
}
\examples{
\dontrun{
epd.connection <- connect_to_epd(host="localhost", database="epd",
                                 user="epdr", password="epdrpw")
epd.1 <- get_entity(1, epd.connection)
check_valid_defaultchron(epd.1)
}
}
