% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr-standardize_functions.R
\name{checkDefaultChronology}
\alias{checkDefaultChronology}
\title{Title TBW}
\usage{
checkDefaultChronology(agedcount)
}
\arguments{
\item{agedcount}{TBW}
}
\value{
TBW
}
\description{
Title TBW
}
\examples{
# TBW
}

