% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr-standardize_functions.R
\name{checkRestriction}
\alias{checkRestriction}
\title{Title TBW}
\usage{
checkRestriction(agedcount)
}
\arguments{
\item{agedcount}{TBW}
}
\value{
TBW
}
\description{
Title TBW
}
\examples{
#  TBW
}

