% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr.R
\name{chrontoCLAM}
\alias{chrontoCLAM}
\title{Reshape a chronology object for CLAM}
\usage{
chrontoCLAM(chronology, C14, chronology_number = NA)
}
\arguments{
\item{chronology}{Chronology object as returned by \code{\link[EPDr:getChronologies]{getChronologies}}.}

\item{C14}{Data frame with C14 data as returned by \code{\link[EPDr:getC14]{getC14}}.}

\item{chronology_number}{Integer indicating the number of the chronology to be used. By default it use the default chronology as in the EPD DDBB.}
}
\value{
Data frame with C14 and noC14 data ready for CLAM
}
\description{
Reshape a chronology object for CLAM
}
\examples{
library(EPDr)
epd.connection <- connectToEPD(host="diegonl.ugr.es", database="epd_ddbb",
                               user="epdr", password="epdrpw")
c14 <- getC14(4, epd.connection)
chron <- getChronologies(4, epd.connection)
chron.clam <- chrontoCLAM(chron, c14)
}

