% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clam.R
\name{clam}
\alias{clam}
\title{Calibrate C14 ages and calculate age-depth models}
\usage{
clam(name = "Example", type = 1, smooth = c(), prob = 0.95,
  its = 1000, wghts = 1, cc = 1, cc1 = "IntCal13.14C",
  cc2 = "Marine13.14C", cc3 = "SHCal13.14C", cc4 = "mixed.14C",
  cc5 = "gluedHemispheres.14C", postbomb = FALSE,
  pb1 = "postbomb_NH1.14C", pb2 = "postbomb_NH2.14C",
  pb3 = "postbomb_NH3.14C", pb4 = "postbomb_SH1_2.14C",
  pb5 = "postbomb_SH3.14C", outliers = c(), ignore = c(),
  youngest = c(), extradates = c(), slump = c(), est = 1,
  calibt = FALSE, mixed.effect = FALSE, dmin = c(), dmax = c(),
  every = 1, yrmin = c(), yrmax = c(), yrsteps = 1, pbsteps = 0.01,
  hpdsteps = 1, BCAD = FALSE, decimals = 0, accrate = 0,
  ageofdepth = c(), depth = "cm", depthseq = c(), depths.file = FALSE,
  thickness = 1, hiatus = c(), remove.reverse = 0.5, times = 5,
  sep = ",", ext = ".csv", runname = c(), storedat = FALSE,
  threshold = 1e-06, proxies = FALSE, revaxes = FALSE, revd = TRUE,
  revyr = TRUE, calhght = 0.3, maxhght = 0.01, mirror = TRUE,
  plotrange = TRUE, bty = "l", mar = c(3.5, 3, 2, 1), mgp = c(2, 1, 0),
  plotpdf = TRUE, plotpng = TRUE, greyscale = c(), yrlab = c(),
  dlab = c(), calcol = rgb(0, 0.5, 0.5, 0.5), C14col = rgb(0, 0, 1, 0.5),
  outcol = "red", outlsize = 1, bestcol = "black", rangecol = rgb(0, 0,
  0, 0.3), slumpcol = grey(0.75), plotname = TRUE, ash = FALSE)
}
\arguments{
\item{name}{The name of a folder with the data as required by clam2.2 as character (e.g. "Example").}

\item{type}{Method type for interpolation: 1 linear interpolation between neighbouring levels ("int", "inter", "interp"); 2 linear or higher polynomial regression ("reg", "regr", "poly" or "polyn", default linear); 3 cubic spline ("spl", "spline"); 4 smooth spline ("sm", "smooth", default smoothing 0.3); 5 locally weighted spline ("loess", "lowess", default smoothing 0.75, cannot extrapolate). Default is 1.}

\item{smooth}{degree of smoothing. Gives polynomial degree for model type 2. Default are: 1 (linear) for type=2, 0.3 for type=4, 0.75 for type=5.}

\item{prob}{Confidence interval (between 0 and 1). Default is 0.95}

\item{its}{Amount of iterations. Default is 1000.}

\item{wghts}{Weights as: 0 no weighting; 1 weighted to calibrated probabilities of sampled calendar years; 2 weighted to (inverse squared) errors of the dates. Default is 1.}

\item{cc}{Calibration curve for C14 dates (1, 2 or 3). Alternatively also 4 and 5, but then cc4 and cc5 curves has to be supplied into the working directory. Default is 1}

\item{cc1}{Calibration curve for northern hemisphere terrestrial C14 dates. This calibration curve is included in the package and no need to be in the working directory. IntCal13.14C}

\item{cc2}{Calibration curve for marine C14 dates. This calibration curve is included in the package and no need to be in the working directory. Marine13.14C}

\item{cc3}{Calibration curve for southern hemisphere C14 dates. This calibration curve is included in the package and no need to be in the working directory. SHCal13.14C}

\item{cc4}{Calibration curve for mixed terrestrial/marine C14 dates. These two calibration curves are included in the package and no need to be in the working directory. mixed.14C}

\item{cc5}{Calibration curve for southern hemisphere C14 dates beyond 11 kcal BP. The new calibration curves extend until 50 kcal BP and, hence, this is not needed any more. However the option is keeped in the package for back compatibility. gluedHemispheres.14C}

\item{postbomb}{Postbomb curve to be used for negative C14 ages. 0 = none. Default is 0.}

\item{pb1}{Postbomb curve for northern hemisphere region 1 postbomb C14 dates: postbomb_NH1.14C}

\item{pb2}{Postbomb curve for northern hemisphere region 2 postbomb C14 dates: postbomb_NH2.14C}

\item{pb3}{Postbomb curve for northern hemisphere region 3 postbomb C14 dates: postbomb_NH3.14C}

\item{pb4}{Postbomb curve for southern hemisphere postbomb region 3 C14 dates: postbomb_SH1-2.14C}

\item{pb5}{Postbomb curve for southern hemisphere postbomb regions 1-2 C14 dates: postbomb_SH3.14C}

\item{outliers}{The number of any dates to be considered outlying, e.g. c(5,6) for the fifth and sixth date counting from the top of a core.}

\item{ignore}{The number of any dates that should be ignored, e.g., c(5,6) for the fifth and sixth date counting from the top of a core.}

\item{youngest}{The age beyond which dates should be truncated}

\item{extradates}{Depths of any additional dates with their files of ages and probabilities.}

\item{slump}{Upper and lower depths of sections of abrupt accumulation that should be excised, e.g., c(600, 550, 120, 100) for two sections of 600-550 and 120-100cm depth.}

\item{est}{????:    1 weighted averages of age-depth model derived ages; 2 midpoints of age-depth model derived ages; 3 midpoints of calibrated ranges; 4 weighted means of calibrated ranges; 5 medians of calibrated distributions; 6 maximum densities of calibrated distributions; 7 midpoints of entire calibrated distributions (with probabilities beyond threshold). Default is 1.}

\item{calibt}{Logical, FALSE by default. Provide two parameters such as c(3,4).}

\item{mixed.effect}{Set to TRUE to activate mixed-effect modelling. Default is FALSE}

\item{dmin}{Minimum depth of age-depth model (e.g., extrapolate)}

\item{dmax}{Maximum depth of age-depth model (e.g., extrapolate)}

\item{every}{Resolution at which (ages for) depths are calculated. Default is 1.}

\item{yrmin}{Minimum of calendar axis of age-depth plot (calculate automatically by default).}

\item{yrmax}{Maximum of calendar axis of age-depth plot (calculated automatically by default).}

\item{yrsteps}{Temporal resolution at which calibrated ages are calculated (in calendar years). Default is 1.}

\item{pbsteps}{Temporal resolution at which postbomb C14 ages are calibrated (in calendar years). Default is 0.01.}

\item{hpdsteps}{Temporal resolution at which highest posterior density ranges are calibrated (in calendar years). Default is 1.}

\item{BCAD}{Logical indicating whether to use BC/AD instead of cal BP. Default is FALSE.}

\item{decimals}{Amount of decimals for rounding. Default is 0.}

\item{accrate}{Give accumulation rates in yr/cm (0) or cm/yr (1). Default is 0.}

\item{ageofdepth}{Calculate age estimates of a specific depth.}

\item{depth}{Depth units. Default is "cm".}

\item{depthseq}{Sequence of depths for which age estimates are to be calculated (default: from dmin to dmax with steps of size every).}

\item{depths.file}{Logical indicating whether to use a file with depths for depthseq. Default is FALSE.}

\item{thickness}{Thickness of the dated samples. Default is 1.}

\item{hiatus}{Depths of any hiatuses, e.g., c(500, 300). Each sub-section must have at least 2 dates (4 for smoothing spline; does not work with loess as it cannot extrapolate).}

\item{remove.reverse}{Proportion of age-models with reversals that can be removed before prompting a warning. Set at FALSE to avoid removing models with reversals. Default is 0.5.}

\item{times}{Half-range of calibration curve used to calibrate dates (multiplication factor for the dates' errors). Default is 5.}

\item{sep}{Separator between the fields of the plain text file containing the dating information. Default is ",".}

\item{ext}{Extension of the file containing the dating information. Default is ".csv".}

\item{runname}{Text to add to the corename for specific runs, e.g., "MyCore_Test1".}

\item{storedat}{Logical indicating whether to store the dates and age-model within R after a clam run. Default is FALSE.}

\item{threshold}{Below which value should probabilities be excluded from calculations. Default is 1e-6.}

\item{proxies}{Logical, default is FALSE. Set to TRUE to plot proxies against age after the run.}

\item{revaxes}{Logical, default is FALSE. Set to TRUE to plot ages on the vertical axis and depth on the horizontal axis.}

\item{revd}{Logical, default is TRUE. Plot depth axis in reverse.}

\item{revyr}{Logical, default is TRUE. Plot age axis in reverse.}

\item{calhght}{Heights of the calibrated distributions in the age-depth plot. Default is 1.}

\item{maxhght}{Maximum height of age probability distributions. Default is 0.1.}

\item{mirror}{Logical, default is TRUE. Plot the age distributions in "mirror" style (above and below depth).}

\item{plotrange}{Logical, default is TRUE. Plot the confidence ranges of the age-model.}

\item{bty}{Default is "1".}

\item{mar}{Vector of 4 elements (c(3.5, 3, 2, 1)). Plot margins (amount of white space along edges of axes 1-4).}

\item{mgp}{Axis text margins (where should titles, labels and tick marks be plotted).}

\item{plotpdf}{Logical, default is TRUE. Produce a pdf file of the age-depth plot}

\item{plotpng}{Logical, default is TRUE. Produce a png file of the age-depth plot}

\item{greyscale}{Produce a grey-scale representation of all age-models (number gives resolution, e.g., 500 bins; will cancel plotting of the confidence intervals). Default is "none".}

\item{yrlab}{Alternative names can be provided. Defaults are "cal BP" or "BC/AD", depending on BCAD value.}

\item{dlab}{Alternative names can be provided. Default is "depth (cm)".}

\item{calcol}{Colour of the calibrated distributions in the age-depth plot. Default is rgb(0,0.5,0.5,0.5) (transparent blue).}

\item{C14col}{Colour of the calibrated ranges of the dates. Default is "blue".}

\item{outcol}{Colour of outlying dates. Default is "red".}

\item{outlsize}{Size of symbols outlying dates. Default is 1.}

\item{bestcol}{Colour of the "best" age-depth model (based on chosen value for est). Default is "black".}

\item{rangecol}{Colour of plotted confidence ranges. Default is rgb(0,0,0,0.3) (transparent grey).}

\item{slumpcol}{Colour of slump. Default is grey(0.75).}

\item{plotname}{Logical indicating whether to print the core name on the graph. Default is TRUE.}

\item{ash}{Logical indicating whether to plot all distributions at the same height. Default is FALSE.}
}
\value{
This function does not return any value or R object, instead it creates a folder structure into the working directory. First it creates a folder called "Cores" in the working directory. In this folder it creates a new folder with the number specified in parameter "name". Inside this folder it creates several .txt and .csv files with the required information to
}
\description{
Calibrate C14 ages and calculate age-depth models to calculate ages at C14 samples and companing files
}
\examples{
# Not run:
# clam("Example")
}
\author{
Maarten Blaauw \email{kbroman@biostat.wisc.edu}, adapted by Diego Nieto-Lugilde \email{dnietolugilde@gmail.com}
}
\references{
\url{http://chrono.qub.ac.uk/blaauw/clam.html}
}
\keyword{age}
\keyword{c14}
\keyword{calibration}
\keyword{clam}

