% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr.R
\name{compareC14Chron4CLAM}
\alias{compareC14Chron4CLAM}
\title{TBW}
\usage{
compareC14Chron4CLAM(chronology, c14, chronology_number = NA,
  include_chron_not_in_c14 = NA, include_c14_not_in_chron = NA,
  use_c14_conf_age = NA, use_c14_conf_depth = NA)
}
\arguments{
\item{chronology}{TBW}

\item{c14}{TBW}

\item{chronology_number}{TBW}

\item{include_chron_not_in_c14}{TBW}

\item{include_c14_not_in_chron}{TBW}

\item{use_c14_conf_age}{TBW}

\item{use_c14_conf_depth}{TBW}
}
\value{
TBW
}
\description{
TBW
}
\examples{
# TBW
}

