% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr.R
\name{core4Clam}
\alias{core4Clam}
\title{Title}
\usage{
core4Clam(core.num, conn, get.dephts = TRUE)
}
\arguments{
\item{core.num}{TBW}

\item{conn}{TBW}

\item{get.dephts}{TBW}
}
\value{
TBW
}
\description{
Title
}
\examples{
# TBW
}

