% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr-epd2clam_functions.R
\name{core4Clam_old}
\alias{core4Clam_old}
\title{Title}
\usage{
core4Clam_old(core_number, connection, get_dephts = TRUE)
}
\arguments{
\item{core_number}{TBW}

\item{connection}{TBW}

\item{get_dephts}{TBW}
}
\value{
TBW
}
\description{
Title
}
\examples{
# TBW
}

