% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr-classes.R
\docType{class}
\name{counts-class}
\alias{counts}
\alias{counts-class}
\title{Title TBW}
\value{
TBW
}
\description{
Title TBW
}
\section{Slots}{

\describe{
\item{\code{e_}}{numeric. TBW}

\item{\code{restriction}}{data.frame. TBW}

\item{\code{counts_type}}{factor. TBW}

\item{\code{counts_processing}}{factor. TBW}

\item{\code{taxa_type}}{factor. TBW}

\item{\code{taxa_processing}}{factor. TBW}

\item{\code{site}}{data.frame. TBW}

\item{\code{taxa_names}}{character. TBW}

\item{\code{taxa_}}{numeric. TBW}

\item{\code{taxa_accepted}}{numeric. TBW}

\item{\code{taxa_mhvar}}{numeric. TBW}

\item{\code{taxa_groupid}}{character. TBW}

\item{\code{sample_}}{numeric. TBW}

\item{\code{sample_label}}{character. TBW}

\item{\code{default_ages}}{numeric. TBW}

\item{\code{depthcm}}{numeric. TBW}

\item{\code{counts}}{data.frame. TBW}
}}
\examples{
# TBW
}

