% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr-standardize_methods.R
\docType{methods}
\name{counts_to_percentage}
\alias{counts_to_percentage}
\alias{counts_to_percentage,epd.entity.df-method}
\alias{counts_to_percentage-method}
\alias{counts_to_percentage,epd.entity-method}
\alias{counts_to_percentage-method}
\title{Calculate counts percentages}
\usage{
counts_to_percentage(x, ..., offset = 1e-04)

\S4method{counts_to_percentage}{epd.entity.df}(x, ..., offset = 1e-04)

\S4method{counts_to_percentage}{epd.entity}(x, ..., offset = 1e-04)
}
\arguments{
\item{x}{\code{\link[EPDr:epd.entity.df]{epd.entity.df}} or
\code{\link[EPDr:epd.entity]{epd.entity}} objects.}

\item{...}{Useless with current methods.}

\item{offset}{Numeric value to avoid dividing by zero if the total of particles is zero for
a particular sample. That may happen when the \code{\link[EPDr:epd.entity.df]{epd.entity.df}}
object has been filtered by taxa groups. For instance, if only trees and shrubs ("TRSH") are 
kept in the \code{@commdf@counts} node some sites in the last glacial maximum could have no
pollen at all, because all vegetation was herbaceos or small shrubs.}
}
\value{
The function returns a modified version of the \code{epd.entity.df} object, in which
values in slot @counts represent percentages instead of raw counts.
}
\description{
This function transforms counts in the slot @commdf@counts in a
\code{\link[EPDr:epd.entity.df]{epd.entity.df}} object to percentages relative to
the total amount of particles in each sample (row).
}
\details{
Details TBW
}
\examples{
# epd.connection <- connectToEPD(host="localhost", database="epd",
# user="epdr", password="epdrpw")
# agedcount <- getAgedCounts(1, epd.connection)
# agedcount.pollen <- filter_taxagroups(agedcount, c("DWAR", "HERB", "LIAN",
# "TRSH", "UPHE", "INUN")) ## All pollen taxa groups
# agedcount.percent <- counts_to_percentage(agedcount.pollen)
# head(agedcount.pollen@counts@counts)
# head(agedcount.percent@counts@counts)

}
