% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr-standardize_methods.R
\docType{methods}
\name{counts_to_percentage}
\alias{counts_to_percentage}
\alias{counts_to_percentage,epd.entity.df-method}
\alias{counts_to_percentage,epd.entity-method}
\title{Calculate counts percentages}
\usage{
counts_to_percentage(x, ..., offset = 1e-04)

\S4method{counts_to_percentage}{epd.entity.df}(x, ..., offset = 1e-04)

\S4method{counts_to_percentage}{epd.entity}(x, ..., offset = 1e-04)
}
\arguments{
\item{x}{epd.entity \code{\link[EPDr]{epd.entity.df}} or 
\code{\link[EPDr]{epd.entity}} objects.}

\item{...}{Useless with current methods.}

\item{offset}{numeric Numeric value to avoid dividing by zero 
if the total of particles is zero for a particular sample. This may 
happen when \code{x} has been filtered by taxa groups. For instance, 
if only trees and shrubs ("TRSH") are kept in the \code{x@commdf@counts} 
slot some sites in the last glacial maximum could have no pollen at all, 
because all vegetation was herbaceos or small shrubs. The effect of the
offset is negligible because it only affect pollen counts of zero. The
result is to indicate that when counts are present, the results is 
numerical zero and not NA. This is relevant to discriminate sites and
times combinations where pollen counts are not available (NA) and not
numerically zero.}
}
\value{
The function returns a \code{\link[EPDr]{epd.entity.df}} object, 
which is a modified version of \code{x} (or \code{entity_to_matrices(x)}
if x is a \code{\link[EPDr]{epd.entity}} object), in which
values in @counts slot represent percentages instead of raw counts.
}
\description{
This function transforms the counts matrix (samples x taxa) in
the @commdf@counts slot of a \code{\link[EPDr]{epd.entity.df}}
object to percentages relative to the total amount of particles 
in each sample (row).
}
\details{
Although the function work with \code{\link[EPDr]{epd.entity.df}} and
\code{\link[EPDr]{epd.entity}} objects, the later are previously 
transformed using \code{\link[EPDr]{entity_to_matrices}} without any
posterior data manipulation and transformation. A lot of entities have
information from different taxonomical groups (terrestrial plants pollen 
and algae or fungi spores). Hence, percentages calculated in this base
may be meaningless and its use is unrecommended.
}
\examples{
# epd.connection <- connect_to_epd(host="localhost", database="epd",
# user="epdr", password="epdrpw")
# epd.1 <- get_entity(1, epd.connection)
# epd.1 <- entity_to_matrices(epd.1)
# epd.1 <- filter_taxagroups(epd.1, c("DWAR", "HERB", "LIAN",
# "TRSH", "UPHE", "INUN")) ## All pollen taxa groups
# epd.1.percent <- counts_to_percentage(epd.1)
# head(epd.1.percent@commdf@counts)
# head(epd.1@commdf@counts)

}
