% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr-classes.R
\docType{class}
\name{datation-class}
\alias{datation}
\alias{datation-class}
\title{TitleTBW}
\value{
TBW
}
\description{
TitleTBW
}
\section{Slots}{

\describe{
\item{\code{e_}}{numeric. TBW}

\item{\code{restriction}}{data.frame. TBW}

\item{\code{entity}}{TBW}

\item{\code{site}}{data.frame. TBW}

\item{\code{postbomb_zone}}{factor. TBW}

\item{\code{chronology}}{chronology. TBW}

\item{\code{c14}}{data.frame. TBW}

\item{\code{events}}{data.frame. TBW}

\item{\code{depths}}{data.frame. TBW}
}}
\examples{
# TBW
}

