% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr-epd2clam_functions.R
\name{datation_epd2clam}
\alias{datation_epd2clam}
\title{Title TBW}
\usage{
datation_epd2clam(x, chronology_number = NA, include_chron_not_in_c14 = NA,
  include_c14_not_in_chron = NA, use_c14_conf_age = NA,
  use_c14_conf_depth = NA, include_depths = T, include_events = NULL)
}
\arguments{
\item{x}{TBW}

\item{chronology_number}{TBW}

\item{include_chron_not_in_c14}{TBW}

\item{include_c14_not_in_chron}{TBW}

\item{use_c14_conf_age}{TBW}

\item{use_c14_conf_depth}{TBW}

\item{include_depths}{TBW}

\item{include_events}{TBW}
}
\value{
TBW
}
\description{
Title TBW
}
\examples{
# TBW
}

