% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr-epd2clam_functions.R, R/EPDr-export_functions.R
\name{datation_export}
\alias{datation_export}
\title{Title TBW}
\usage{
datation_export(datation, format = c("clam", "bacon"),
  chronology_number = NA, include_chron_not_in_c14 = NA,
  include_c14_not_in_chron = NA, use_c14_conf_age = NA,
  use_c14_conf_depth = NA, include_depths = T, include_events = NULL)

datation_export(datation, format = c("clam", "bacon"),
  chronology_number = NA, include_chron_not_in_c14 = NA,
  include_c14_not_in_chron = NA, use_c14_conf_age = NA,
  use_c14_conf_depth = NA, include_depths = T, include_events = NULL)
}
\arguments{
\item{datation}{TBW}

\item{format}{TBW}

\item{chronology_number}{TBW}

\item{include_chron_not_in_c14}{TBW}

\item{include_c14_not_in_chron}{TBW}

\item{use_c14_conf_age}{TBW}

\item{use_c14_conf_depth}{TBW}

\item{include_depths}{TBW}

\item{include_events}{TBW}

\item{datation}{TBW}

\item{format}{TBW}

\item{chronology_number}{TBW}

\item{include_chron_not_in_c14}{TBW}

\item{include_c14_not_in_chron}{TBW}

\item{use_c14_conf_age}{TBW}

\item{use_c14_conf_depth}{TBW}

\item{include_depths}{TBW}

\item{include_events}{TBW}
}
\value{
TBW

TBW
}
\description{
Title TBW

Title TBW
}
\examples{
# TBW
# TBW
}

