% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr-epd2clam_functions.R
\name{depths_epd2clam}
\alias{depths_epd2clam}
\title{Reshape depths data to CLAM format}
\usage{
depths_epd2clam(depths)
}
\arguments{
\item{depths}{Data frame with at least a column called depthcm.}
}
\value{
Vector with depths in ascending order.
}
\description{
This function takes depths data, as returned by \code{\link[EPDr:getDepths]{getDepths}}, to comply with CLAM format.
}
\examples{
epd.connection <- connectToEPD(database="epd_ddbb", user="epdr",
                                 password="epdrpw", host="diegonl.ugr.es")
depths.1 <- getDepths(1, epd.connection)
depths_epd2clam(depths.1)
disconnectFromEPD(connection=epd.connection)
}

