% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr-export_functions.R
\name{depths_export}
\alias{depths_export}
\title{Reshape depths data to CLAM or BACON format}
\usage{
depths_export(depths)
}
\arguments{
\item{depths}{Data frame with at least a column called depthcm.}
}
\value{
Vector with depths in ascending order.
}
\description{
This function takes depths data, as returned by \code{\link[EPDr:getDepths]{getDepths}}, to comply with CLAM or BACON format.
}
\examples{
#epd.connection <- connectToEPD(database="epd_ddbb", user="epdr",
#                                 password="epdrpw", host="localhost")
#depths.1 <- getDepths(1, epd.connection)
#depths_epd2clam(depths.1)
#disconnectFromEPD(connection=epd.connection)
}

