% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDPr-DDBB_connection.R
\name{disconnect_from_epd}
\alias{disconnect_from_epd}
\title{Disconnect a connection to a EPD database}
\usage{
disconnect_from_epd(connection = NULL)
}
\arguments{
\item{connection}{PostgreSQLConnection The connection object created with 
\code{\link[EPDr]{connect_to_epd}} to stablish the connection}
}
\value{
NULL It just disconnects from the EPD DDBB server and modifies 
the connection object to reflect the new status.
}
\description{
\code{disconnect_from_epd} turns down a connection to a EPD DDBB server.
}
\examples{
# epd.connection <- connect_to_epd(database="epd", user="epdr",
#                                 password="epdrpw", host="localhost")
# disconnect_from_epd(connection=epd.connection)
# epd.connection
}
