% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr-search_functions.R
\name{e_by_countries}
\alias{e_by_countries}
\title{Title}
\usage{
e_by_countries(countries, connection, regions = NULL)
}
\arguments{
\item{countries}{}

\item{connection}{}

\item{regions}{}
}
\description{
Title
}

