% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr-search_functions.R
\name{e_by_publ}
\alias{e_by_publ}
\title{Title TBW}
\usage{
e_by_publ(publ, connection)
}
\arguments{
\item{publ}{TBW}

\item{connection}{TBW}
}
\value{
TBW
}
\description{
Title TBW
}
\examples{
# TBW
}

