% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr-search_functions.R
\name{e_by_restriction}
\alias{e_by_restriction}
\title{Title TBW}
\usage{
e_by_restriction(restriction, connection)
}
\arguments{
\item{restriction}{TBW}

\item{connection}{TBW}
}
\value{
TBW
}
\description{
Title TBW
}
\examples{
#  TBW
}

