% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr-search_functions.R
\name{e_by_site}
\alias{e_by_site}
\title{Title TBW}
\usage{
e_by_site(site, connection)
}
\arguments{
\item{site}{TBW}

\item{connection}{TBW}
}
\value{
TBW
}
\description{
Title TBW
}
\examples{
# TBW
}

