% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr-search_functions.R
\name{e_by_worker}
\alias{e_by_worker}
\title{Title TBW}
\usage{
e_by_worker(lastname, connection, firstname = NULL, initials = NULL)
}
\arguments{
\item{lastname}{TBW}

\item{connection}{TBW}

\item{firstname}{TBW}

\item{initials}{TBW}
}
\value{
TBW
}
\description{
Title TBW
}
\examples{
# TBW
}

