% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr-classes.R
\docType{class}
\name{entity-class}
\alias{entity-class}
\alias{entity}
\title{Class for descriptions of an entity}
\description{
Class \code{entity} stores all summary and description information relative 
to a specific entity of the European Pollen Database (EPD). This object
is created by \code{\link[EPDr]{get_ent}} function.
}
\section{Slots}{

\describe{
\item{\code{entity}}{data.frame. Information describing some characteristics of
the entity.}

\item{\code{pentity}}{data.frame. Description of contact person and restriction on
the use of data from this entity.}

\item{\code{contact}}{data.frame. Details on the contact person.}

\item{\code{coredrive}}{data.frame. Description of the coredrive used to sample
the entity.}

\item{\code{descr}}{data.frame. This table provide a longer explanation for the
descriptor specified in \code{@entity} slot.}

\item{\code{lithology}}{data.frame. Information about the lithology found in the
entity.}

\item{\code{loi}}{data.frame. Detailed information for loss-on-ignition data for
that particular entity.}

\item{\code{section}}{data.frame. Details on the sections used to extract the
entity if is a core.}

\item{\code{publent}}{data.frame. Linking information between entity id and
publications (\code{@publ}).}

\item{\code{publ}}{data.frame PUBL data where the event data were published.}
}}

