% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr-standardize_methods.R
\docType{methods}
\name{entity_to_matrices}
\alias{entity_to_matrices}
\alias{entity_to_matrices,epd.entity-method}
\alias{entity_to_matrices-method}
\title{entity_to_matrices title TBW}
\usage{
entity_to_matrices(epd.entity)

\S4method{entity_to_matrices}{epd.entity}(epd.entity)
}
\arguments{
\item{epd.entity}{epd.entity object. TBW}
}
\value{
TBW
}
\description{
entity_to_matrices description TBW
}
\details{
entity_to_matrices details TBW
}
\examples{
# TBW
}
