% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr-classes.R
\docType{class}
\name{epd.entity-class}
\alias{epd.entity-class}
\alias{epd.entity}
\title{Class for epd.entity objects}
\description{
This object stores all the information available for a particular 
entity in the European Pollen Database (EPD). It has some additional 
information (e.g., in \code{@postbombzone} or \code{@isingiesecke} slots).
Other slots store handy information for simplicity (e.g.,
\code{@defaultchron} or \code{@numberofchron} slots). The rest of 
slots are the raw information as extracted from the EPD, and
stored in the \code{link[EPDr]{entity}}, \code{link[EPDr]{site}}, 
\code{link[EPDr]{geochron}}, \code{link[EPDr]{chron}}, and 
\code{link[EPDr]{samples}} classes.
}
\details{
All data in this object can be changed, but the intention
of the object is to provide a copy of all the tables in the EPD
with information regarding a particular entity. When an entity
has no data in any particular table, the table is created empty.
}
\section{Slots}{

\describe{
\item{\code{e_}}{numeric. Number indicating the entity number (e_) of the
stored entity.}

\item{\code{postbombzone}}{factor. It indicates the code of the postbomb zone of
the world where the entity is located. This is not used yet, but
might be handy for further implementations for datations with CLAM
or BACON.}

\item{\code{numberofchron}}{numeric. The number of chronologies available
for the data in that entity.}

\item{\code{isingiesecke}}{logical. Indicating whether this entity was
revised in Giesecke et al. (2013) and, hence, additional datations
are available.}

\item{\code{defaultchron}}{numeric. The number of the default chronology
according to the EPD.}

\item{\code{entity}}{entity. Entity object for the entity.}

\item{\code{site}}{site. Site object for the entity.}

\item{\code{geochron}}{geochron. Geochron object for the entity.}

\item{\code{chron}}{chron. Chron object for the entity.}

\item{\code{samples}}{samples. Samples object for the entity.}
}}

\references{
Giesecke, T., Davis, B., Brewer, S., Finsinger, W., 
Wolters, S., Blaaw, M., de Beaulieu, J.L., Binney, H., Fyfe, R.M.,
Gaillard, M.J., Gil-Romera, G., van der Knaap, W.O. Kunes, P.,
Kuhl, N., van Leeuwen, J.F.N, Leydet, M., Lotter, A.F., Ortu, E.,
Semmler, M., and Bradshaw, R.H.W (2013). Towards mapping the late
Quaternary vegetation change of Europe. Vegetation History and
Archaeobotany, 23, 75-86.
}
