% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr-classes.R
\docType{class}
\name{epd.entity.df-class}
\alias{epd.entity.df-class}
\alias{epd.entity.df}
\title{Class for epd.entity.df objects}
\description{
This is an expansion of \code{\link[EPDr]{epd.entity}} class,
in which certain data have been reformated and included in the object. Hence,
the first @slots are the same as in a
\code{\link[EPDr]{epd.entity}}. Then, additional data
include four @slots for information on the processing and
transformation of the data, and for @slots for reformated data.
}
\section{Slots}{

\describe{
\item{\code{e_}}{numeric. Number indicating the entity number (e_) of the
stored entity. Same as in a \code{\link[EPDr]{epd.entity}} 
object.}

\item{\code{postbombzone}}{factor. It indicate the code of the postbombzone of
the world where the entity is located. This is not used yet, but
might be handy for further implementations for datations with CLAM
or BACON. Same as in a \code{\link[EPDr]{epd.entity}} 
object.}

\item{\code{numberofchron}}{numeric. The number of chronologies available
for the data in that entity.Same as in a 
\code{\link[EPDr]{epd.entity}} object.}

\item{\code{isingiesecke}}{logical. Indicating whether this entity was
revised in Giesecke et al. (2013) and, hence, additional datations
are available. Same as in a \code{\link[EPDr]{epd.entity}} 
object.}

\item{\code{defaultchron}}{numeric. The number of the default chronology
according to the EPD. Same as in a \code{\link[EPDr]{epd.entity}} 
object.}

\item{\code{entity}}{entity. Entity object for the entity. Same as in 
a \code{\link[EPDr]{epd.entity}} object.}

\item{\code{site}}{site. Site object for the entity. Same as in a 
\code{\link[EPDr]{epd.entity}} object.}

\item{\code{geochron}}{geochron. Geochron object for the entity. Same as in 
a \code{\link[EPDr]{epd.entity}} object.}

\item{\code{chron}}{chron. Chron object for the entity. Same as in 
a \code{\link[EPDr]{epd.entity}} object.}

\item{\code{samples}}{samples. Samples object for the entity. Same as in 
a \code{\link[EPDr]{epd.entity}} object.}

\item{\code{countstype}}{factor. Indicating if the counts in @slot commdf
are in raw counts or percentages.}

\item{\code{countsprocessing}}{factor. Indicating whether the 
counts are those counted (or percentaged) for the samples, or
have been interpolated or intervaled for particular time periods
(or intervals). See \code{\link[EPDr]{interpolate_counts}} or
\code{\link[EPDr]{intervals_counts}}.}

\item{\code{taxatype}}{factor. Indicating whether the taxa are the
same as in the original database, or they have been changed to
accepted taxa or higher taxonomical level.}

\item{\code{taxaprocessing}}{factor. Indicating whether the taxa in @slot 
commdf are those originally present in the database, or
they have been expanded or filtered.}

\item{\code{samplesdf}}{samplesdf. Samples in vector format.}

\item{\code{agesdf}}{agesdf. Ages in vector and table format.}

\item{\code{commdf}}{commdf. Counts of biological particles in vector and
table format.}

\item{\code{nopodf}}{nopodf. Counts of no-biological particles in vector
and table format.}
}}

