% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr-clam_functions.R
\name{epdrClam}
\alias{epdrClam}
\title{CLAM calibration with automatic postbomb zone selection}
\usage{
epdrClam(core_number, connection, ...)
}
\arguments{
\item{core_number}{Integer indicating the number of the core (entity) to be calibrated with CLAM}

\item{connection}{The connection to the EPD to get the geographical position of the core}

\item{...}{Any argument for classical \code{\link[EPDr:clam]{clam}} function.}
}
\value{
The same set of
}
\description{
Recalibrate using CLAM but automagically capturing the coordinates of the core to get the right postbomb zone
 to be used in the calibration. Similar to the \code{\link[EPDr:clam]{clam}} function, \code{epdrCLAM} read calibration files
 from the following folder structure: \code{Cores/code_number}.
}
\examples{
#TBW
}

