% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr.R
\name{event2clam}
\alias{event2clam}
\title{Reshape event data to CLAM format}
\usage{
event2clam(event)
}
\arguments{
\item{event}{Data frame with event data for a particular core (entity) in the EPD DDBB.}
}
\value{
Data frame with event data in CLAM format.
}
\description{
This function takes event data to fit into a new table that comply with CLAM format.
}
\examples{
epd.connection <- connectToEPD(host="diegonl.ugr.es", database="epd_ddbb",
                               user="epdr", password="epdrpw")
synevent <- dbGetQuery(epd.connection, "SELECT * FROM synevent WHERE e_ = 51;")
event <- dbGetQuery(epd.connection, paste("SELECT * FROM event WHERE event_ = ",
                    synevent$event_, ";", sep=""))
event <- merge(synevent, event, by="event_")
event.clam <- event2clam(event)
}

