% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr-epd2clam_functions.R, R/EPDr-export_functions.R
\name{events_export}
\alias{events_export}
\title{Reshape events data to CLAM or BACON format}
\usage{
events_export(event, format = c("clam", "bacon"))

events_export(event, format = c("clam", "bacon"))
}
\arguments{
\item{event}{Data frame with event data for a particular core (entity) in the EPD DDBB.}

\item{format}{Character string indicating whether to export to "clam" or to "bacon" format.}

\item{event}{Data frame with event data for a particular core (entity) in the EPD DDBB.}

\item{format}{Character string indicating whether to export to "clam" or to "bacon" format.}
}
\value{
Data frame with event data in CLAM or BACON format.

Data frame with event data in CLAM or BACON format.
}
\description{
This function takes event data to fit into a new table that comply with CLAM or BACON format.

This function takes event data to fit into a new table that comply with CLAM or BACON format.
}
\examples{
#epd.connection <- connectToEPD(host="localhost", database="epd_ddbb",
#                               user="epdr", password="epdrpw")
#event <- getEvents(51, epd.connection)
#event.clam <- events_export(event, format="clam")
#event.bacon <- events_export(event, format="bacon")
#epd.connection <- connectToEPD(host="localhost", database="epd_ddbb",
#                               user="epdr", password="epdrpw")
#event <- getEvents(51, epd.connection)
#event.clam <- events_export(event, format="clam")
#event.bacon <- events_export(event, format="bacon")
}

