% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr-standardize_functions.R
\name{expandTaxonomy}
\alias{expandTaxonomy}
\title{Title TBW}
\usage{
expandTaxonomy(count, taxa_list, epd_taxonomy)
}
\arguments{
\item{count}{TBW}

\item{taxa_list}{TBW}

\item{epd_taxonomy}{TBW}
}
\value{
TBW
}
\description{
Title TBW
}
\examples{
# TBW
}

