% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr-export_methods.R
\docType{methods}
\name{export_agebasis}
\alias{export_agebasis}
\alias{export_agebasis,character,data.frame-method}
\alias{export_agebasis-method}
\alias{export_agebasis,character,epd.entity-method}
\alias{export_agebasis-method}
\title{Reshape agebasis table to CLAM or BACON format}
\usage{
export_agebasis(format, x)

\S4method{export_agebasis}{character,data.frame}(format, x)

\S4method{export_agebasis}{character,epd.entity}(format, x)
}
\arguments{
\item{format}{Character string indicating whether to export to "clam" or to "bacon" format.}

\item{x}{Data frame or \code{\link[EPDr:epd.entity]{epd.entity}} with agebasis data as those
in a agebasis table returned by \code{\link[EPDr:.get_agebasis]{.get_agebasis}} or
\code{\link[EPDr:get_entity]{get_entity}}.}
}
\value{
Data frame with no-C14 data in CLAM or BACON format. This data frame can be easily
combined with C14 data from \code{\link[EPDr:export_c14]{export_c14}} using \code{rbind}.
}
\description{
This function takes agebasis data, used in previous chronologies, to fit
into a new table that comply with CLAM or BACON format.
}
\examples{
#library(EPDr)
#epd.connection <- connect_to_epd(host = "localhost", database = "epd",
#                               user = "epdr", password = "epdrpw")
#c14 <- getC14_old(400, epd.connection)
#c14.clam <- export_c14(c14, "clam")
#c14.bacon <- export_c14(c14, "bacon")
#chron <- getChronology(400, epd.connection)
#noc14.clam <- export_agebasis(chron@agebasis, "clam")
#noc14.bacon <- export_agebasis(chron@agebasis, "bacon")
#all.clam <- rbind(c14.clam, noc14.clam)
#all.bacon <- rbind(c14.bacon, noc14.bacon)
}
\references{
\url{http://www.chrono.qub.ac.uk/blaauw/clam.html}

\url{http://chrono.qub.ac.uk/blaauw/bacon.html}
}
