% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr-export_methods.R
\docType{methods}
\name{export_agebasis}
\alias{export_agebasis}
\alias{export_agebasis,character,data.frame-method}
\alias{export_agebasis,character,epd.entity-method}
\title{Reshape agebasis table to CLAM or BACON format}
\usage{
export_agebasis(format, x)

\S4method{export_agebasis}{character,data.frame}(format = c("clam", "bacon"),
  x)

\S4method{export_agebasis}{character,epd.entity}(format, x)
}
\arguments{
\item{format}{character Character string indicating whether to export 
to "clam" or to "bacon" format.}

\item{x}{data.frame Data frame or \code{\link[EPDr]{epd.entity}} with
agebasis data as those in an agebasis table returned by 
\code{\link[EPDr]{.get_agebasis}} or \code{\link[EPDr]{get_entity}}.}
}
\value{
Data frame with no-C14 data in CLAM or BACON format. This 
data frame can be easily combined with C14 data from
\code{\link[EPDr]{export_c14}} using \code{rbind}.
}
\description{
This function takes agebasis data, as those from 
\code{\link[EPDr]{.get_agebasis}} or from an \code{\link[EPDr]{epd.entity}},
used to calibrate chronologies of that entity in the EPD, and modifies
them to fit into a new table that complies with CLAM or BACON format.
}
\examples{
# epd.connection <- connect_to_epd(host = "localhost", database = "epd",
#                               user = "epdr", password = "epdrpw")
# epd.400 <- get_entity(400, epd.connection)
# epd.400.c14 <- export_c14("clam", epd.400)
# epd.400.ageb <- export_agebasis("clam", epd.400)
# rbind(epd.400.c14, epd.400.ageb)
# epd.400.c14 <- export_c14("bacon", epd.400)
# epd.400.ageb <- export_agebasis("bacon", epd.400)
# rbind(epd.400.c14, epd.400.ageb)
}
\references{
\url{http://www.chrono.qub.ac.uk/blaauw/clam.html}

\url{http://chrono.qub.ac.uk/blaauw/bacon.html}
}
