% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr-export_methods.R
\docType{methods}
\name{export_depths}
\alias{export_depths}
\alias{export_depths,data.frame-method}
\alias{export_depths-method}
\alias{export_depths,epd.entity-method}
\alias{export_depths-method}
\title{Reshape depths of biological samples to CLAM or BACON format}
\usage{
export_depths(x)

\S4method{export_depths}{data.frame}(x)

\S4method{export_depths}{epd.entity}(x)
}
\arguments{
\item{x}{Data frame with at least a column called depthcm or \code{\link[EPDr:epd.entity]{epd.entity}} object to extract this information from.}
}
\value{
Vector with depths in ascending order.
}
\description{
This function takes depths data, as returned by \code{\link[EPDr:.get_psamples]{.get_psamples}},
to comply with CLAM or BACON format.
}
\examples{
#epd.connection <- connect_to_epd(database = "epd", user = "epdr",
#                                 password = "epdrpw", host = "localhost")
#depths.1 <- getPSamples(1, epd.connection)
#export_depths(depths.1)
#disconnectFromEPD(connection = epd.connection)
}
\references{
\url{http://www.chrono.qub.ac.uk/blaauw/clam.html}

\url{http://chrono.qub.ac.uk/blaauw/bacon.html}
}
