% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr-export_methods.R
\docType{methods}
\name{export_depths}
\alias{export_depths}
\alias{export_depths,data.frame-method}
\alias{export_depths,epd.entity-method}
\title{Reshape depths of biological samples to CLAM or BACON format}
\usage{
export_depths(x)

\S4method{export_depths}{data.frame}(x)

\S4method{export_depths}{epd.entity}(x)
}
\arguments{
\item{x}{data.frame Data frame with at least a column called \code{depthcm}
or \code{\link[EPDr]{epd.entity-class}} object to extract this information from.}
}
\value{
Vector with depths in ascending order.
}
\description{
This function takes depths data, as returned by 
\code{\link[EPDr]{get_samples}} or extracted from a 
\code{\link[EPDr]{epd.entity-class}} object, and transforms them
to comply with CLAM or BACON format.
}
\examples{
\dontrun{
epd.connection <- connect_to_epd(host = "localhost", database = "epd",
                              user = "epdr", password = "epdrpw")
depths.1 <- .get_psamples(1, epd.connection)
export_depths(depths.1)
epd.1 <- get_entity(1, epd.connection)
export_depths(epd.1)
}
}
\references{
\url{http://www.chrono.qub.ac.uk/blaauw/clam.html}

\url{http://chrono.qub.ac.uk/blaauw/bacon.html}
}
