% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr-export_methods.R
\docType{methods}
\name{export_entity}
\alias{export_entity}
\alias{export_entity,character,epd.entity-method}
\title{Reshape epd.entity objects to CLAM or BACON format}
\usage{
export_entity(format, x, chronology = NULL, incl_chron_not_in_c14 = NULL,
  incl_c14_not_in_chron = NULL, use_c14_conf_age = NULL,
  use_c14_conf_depth = NULL, include_depths = TRUE, incl_events = NULL)

\S4method{export_entity}{character,epd.entity}(format, x, chronology = NULL,
  incl_chron_not_in_c14 = NULL, incl_c14_not_in_chron = NULL,
  use_c14_conf_age = NULL, use_c14_conf_depth = NULL,
  include_depths = TRUE, incl_events = NULL)
}
\arguments{
\item{format}{character Character string indicating whether to export to "clam" 
or to "bacon" format.}

\item{x}{epd.entity \code{\link[EPDr]{epd.entity}} or 
\code{\link[EPDr]{epd.entity.df}} objects from which all information 
will be extracted to compose the clam or bacon table and file.}

\item{chronology}{numeric Number indicating the chronology from which 
to extract the agebasis for the "clam" or "bacon" file. If 
unspecified, the function use the default chronology according to 
the EPD for that particular entity.}

\item{incl_chron_not_in_c14}{logical Logical value indicating whether the 
function should include in the result agebasis present in the specified 
(or the default) chronology (TRUE) but not in the C14 table, or should 
not include them (FALSE).}

\item{incl_c14_not_in_chron}{logical Logical value indicating whether the 
function should include in the result agebasis present in the C14 
data but not present in the specified (or default) chronology (TRUE), 
or should not include them (FALSE).}

\item{use_c14_conf_age}{logical Logical value indicating whether the 
function should use data from C14 table when there are conflicting 
ages between the C14 table or the chronology (TRUE), or it should 
take data from the chronology instead (FALSE).}

\item{use_c14_conf_depth}{logical Logical value indicating whether the function 
should use data from C14 table when there are conflicting depths 
between the C14 table or the chronology (TRUE), or it should take data 
from the chronology instead (FALSE).}

\item{include_depths}{logical Logical value indicating whether depths of pollen 
samples should be exported too. The default value is TRUE. This is 
helpful because running CLAM or BACON with depths calculate the 
calibrated ages for those samples in the same step.}

\item{incl_events}{logical Logical value indicating whether events information 
should be included (TRUE), or not (FALSE).}
}
\value{
Data frame with specific format for "CLAM" or "BACON" age-depth
modelling softwares. CLAM format has 7 columns: \code{$ID}, 
\code{$C14_age}, \code{$cal_age}, \code{$error}, \code{$reservoir}, 
\code{$depth}, and \code{$thickness}. \code{$ID} is the code 
of the radiocarbon (C14) samples. \code{$C14_age} is the radiocarbon (C14)
dates. \code{$error} is the error estimated in the radiocarbon (C14) 
datation of the samples. \code{$reservoir} is to specify if the 
samples have marine reservoir effects. \code{$depth} is the depth in
cm of the radiocarbon samples. \code{$thickness} is the tickness of 
the radiocarbon samples. BACON format has 4 columns: \code{$labID},
\code{$age}, \code{$error}, and \code{$depth}. \code{$labID} is the 
code of the radiocarbon (C14) samples. \code{$age} is the radiocarbon 
(C14) date. \code{$error} is the error estimated in the radiocarbon 
(C14) datation for the samples. \code{$depth} is the depth in cm of 
the radiocarbon samples. The function also produce the files required
by CLAM or BACON in the working directory according to this folder 
structure: \code{~/{format}/Cores/{entity_number}/}
The function creates a file \code{.csv} (\code{{entity_number}.csv}) with 
the radiocarbon data and two \code{.txt} (\code{{entity_number}_depths.txt}
and \code{{entity_number}_depthsID.txt}). The last one is not used by
CLAM or BACON but it might be useful to follow track of samples depths.
}
\description{
This function takes \code{\link[EPDr]{epd.entity}} or
\code{\link[EPDr]{epd.entity.df}} objects, as those returned by
\code{\link[EPDr]{get_entity}} or \code{\link[EPDr]{entity_to_matrices}},
extracts datation information and reshapes it to complies with
CLAM or BACON format.
}
\details{
The function has an interactive implementation. If some of the parameters
\code{incl_chron_not_in_c14}, \code{incl_c14_not_in_chron}, 
\code{use_c14_conf_age}, and/or \code{use_c14_conf_depth} are not 
specified, the function will check the data and ask the user how 
it should proceed regarding some conflicts in the data between C14 data
and agebasis tables.
}
\examples{
\dontrun{
epd.connection <- connect_to_epd(host = "localhost", database = "epd",
                              user = "epdr", password = "epdrpw")
epd.400 <- get_entity(400, epd.connection)
export_entity("clam", epd.400) # Also check new folders in your working directory
export_entity("bacon", epd.400) # Also check new folders in your working directory
epd.1 <- get_entity(1, epd.connection)
export_entity("clam", epd.1)
export_entity("bacon", epd.1)
}
}
\references{
\url{http://www.chrono.qub.ac.uk/blaauw/clam.html}

\url{http://chrono.qub.ac.uk/blaauw/bacon.html}
}
