% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr-export_methods.R
\docType{methods}
\name{export_events}
\alias{export_events}
\alias{export_events,character,data.frame,data.frame-method}
\alias{export_events-method}
\alias{export_events,character,epd.entity,missing-method}
\alias{export_events-method}
\title{Reshape events data to CLAM or BACON format}
\usage{
export_events(format, x, y)

\S4method{export_events}{character,data.frame,data.frame}(format, x, y)

\S4method{export_events}{character,epd.entity,missing}(format, x, y)
}
\arguments{
\item{format}{Character string indicating whether to export to "clam" or to "bacon" format.}

\item{x}{Data frame or \code{\link[EPDr:epd.entity]{epd.entity}} object with event data as
those extracted from \code{\link[EPDr:.get_synevent]{.get_synevent}} or \code{\link[EPDr:get_entity]{get_entity}}}

\item{y}{Data frame with event data as those extracted from \code{\link[EPDr:.get_event]{.get_event}}.}
}
\value{
Data frame with event data in CLAM or BACON format.
}
\description{
This function takes event data to fit into a new table that comply with CLAM or BACON format.
}
\examples{
#epd.connection <- connect_to_epd(host = "localhost", database = "epd",
#                               user = "epdr", password = "epdrpw")
#event <- getEvents(51, epd.connection)
#event.clam <- export_events(event, format = "clam")
#event.bacon <- export_events(event, format = "bacon")
}
\references{
\url{http://www.chrono.qub.ac.uk/blaauw/clam.html}

\url{http://chrono.qub.ac.uk/blaauw/bacon.html}
}
