% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr-export_methods.R
\docType{methods}
\name{export_events}
\alias{export_events}
\alias{export_events,character,data.frame,data.frame-method}
\alias{export_events,character,epd.entity,missing-method}
\title{Reshape events data to CLAM or BACON format}
\usage{
export_events(format, x, y)

\S4method{export_events}{character,data.frame,data.frame}(format, x, y)

\S4method{export_events}{character,epd.entity,missing}(format, x, y)
}
\arguments{
\item{format}{character Character string indicating whether to export to "clam" or to "bacon" format.}

\item{x}{data.frame Data frame or \code{\link[EPDr]{epd.entity}} object with event data as
those extracted from \code{\link[EPDr]{.get_synevent}} or \code{\link[EPDr]{get_entity}}}

\item{y}{data.frame Data frame with event data as those extracted from \code{\link[EPDr]{.get_event}}.}
}
\value{
Data frame with event data in CLAM or BACON format.
}
\description{
This function takes event data, as those queried by 
\code{\link[EPDr]{.get_synevent}} or stored in a 
\code{\link[EPDr]{epd.entity}} object, and modifies them to fit into a new 
table that complies with CLAM or BACON format.
}
\details{
If \code{x} is an \code{\link[EPDr]{epd.entity}} or an 
\code{\link[EPDr]{epd.entity.df}} object, \code{y} is not used.
}
\examples{
\dontrun{
epd.connection <- connect_to_epd(host = "localhost", database = "epd",
                              user = "epdr", password = "epdrpw")
epd.51 <- get_entity(51, epd.connection)
export_events("clam", epd.51)
export_events("clam", .get_synevent(51, epd.connection), .get_event(26, epd.connection))
export_events("bacon", epd.51)
export_events("bacon", .get_synevent(51, epd.connection), .get_event(26, epd.connection))
}
}
\references{
\url{http://www.chrono.qub.ac.uk/blaauw/clam.html}

\url{http://chrono.qub.ac.uk/blaauw/bacon.html}
}
