% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr-export_functions.R
\name{exportAgebasis}
\alias{exportAgebasis}
\title{Reshape no-C14 data to CLAM or BACON format}
\usage{
exportAgebasis(agebasis, format = c("clam", "bacon"))
}
\arguments{
\item{agebasis}{Data frame with no-C14 data as those in a chronology list returned by
\code{\link[EPDr:getChronology]{getChronology}}.}

\item{format}{Character string indicating whether to export to "clam" or to "bacon" format.}
}
\value{
Data frame with no-C14 data in CLAM or BACON format. This data frame can be easily combined with C14
data from \code{\link[EPDr:exportC14]{exportC14}} using \code{rbind}.
}
\description{
This function takes no-C14 data, that can be extracted from a chronology object, to fit into a new table that
comply with CLAM or BACON format.
}
\examples{
#library(EPDr)
#epd.connection <- connectToEPD(host="localhost", database="epd",
#                               user="epdr", password="epdrpw")
#c14 <- getC14_old(400, epd.connection)
#c14.clam <- exportC14(c14, "clam")
#c14.bacon <- exportC14(c14, "bacon")
#chron <- getChronology(400, epd.connection)
#noc14.clam <- exportAgebasis(chron@agebasis, "clam")
#noc14.bacon <- exportAgebasis(chron@agebasis, "bacon")
#all.clam <- rbind(c14.clam, noc14.clam)
#all.bacon <- rbind(c14.bacon, noc14.bacon)

}
\references{
\url{http://www.chrono.qub.ac.uk/blaauw/clam.html}

\url{http://chrono.qub.ac.uk/blaauw/bacon.html}
}

