% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr-export_functions.R
\name{exportC14}
\alias{exportC14}
\title{Reshape C14 data to CLAM or BACON format}
\usage{
exportC14(C14, format = c("clam", "bacon"))
}
\arguments{
\item{C14}{Data frame with C14 data as those extracted from \code{\link[EPDr:getC14]{getC14}}.}

\item{format}{Character string indicating whether to export to "clam" or to "bacon" format.}
}
\value{
Data frame with C14 data in CLAM or BACON format.
}
\description{
This function takes C14 data, as those extracted from \code{\link[EPDr:getC14]{getC14}}, to fit into a new table that comply with CLAM or BACON format.
}
\examples{
# Not run
# library(EPDr)
# epd.connection <- connectToEPD(host="localhost", database="epd",
#                               user="epdr", password="epdrpw")
# c14 <- getC14(400, epd.connection)
# c14.clam <- exportC14(c14, "clam")
# c14.bacon <- exportC14(c14, "bacon")

}
\references{
\url{http://www.chrono.qub.ac.uk/blaauw/clam.html}

\url{http://chrono.qub.ac.uk/blaauw/bacon.html}
}

