% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr-export_functions.R
\name{exportDatation}
\alias{exportDatation}
\title{Reshape datation data to CLAM or BACON format}
\usage{
exportDatation(datation, format = "clam", chronology_number = NA,
  include_chron_not_in_c14 = NA, include_c14_not_in_chron = NA,
  use_c14_conf_age = NA, use_c14_conf_depth = NA, include_depths = T,
  include_events = NULL)
}
\arguments{
\item{datation}{\code{\link[EPDr:datation]{datation}} object from which all information will be extracted to compose the clam or bacon file.}

\item{format}{Character string indicating whether to export to "clam" or to "bacon" format.}

\item{chronology_number}{Number indicating the chronology from which to extract the agebasis for the "clam" or "bacon" file. If unspecified, the function use the default chronology according to the EPD for that particular entity.}

\item{include_chron_not_in_c14}{Logical value indicating whether the function should include in the result agebasis present in the specified (or the default) chronology (TRUE) but not in the C14 table, or should not include them (FALSE).}

\item{include_c14_not_in_chron}{Logical value indicating whether the function should include in the result agebasis present in the C14 data but not present in the specified (or default) chronology (TRUE), or should not include them (FALSE).}

\item{use_c14_conf_age}{Logical value indicating whether the function should use data from C14 table when there are conflicting ages between the C14 table or the chronology (TRUE), or it should take data from the chronology instead (FALSE).}

\item{use_c14_conf_depth}{Logical value indicating whether the function should use data from C14 table when there are conflicting depths between the C14 table or the chronology (TRUE), or it should take data from the chronology instead (FALSE).}

\item{include_depths}{Logical value indicating whether depths of biological samples should be exported too. The default value is TRUE. This is helpful because running CLAM or BACON with depths calculate the calibrated ages for those sample in the same step.}

\item{include_events}{Logical value indicating whether events information should be included (TRUE), or not (FALSE).}
}
\value{
Data frame with specific format for "CLAM" or "BACON" age-depth modelling softwares.
CLAM format has 7 columns: lab_ID, C14_age, cal_age, error, reservoir, depth, and thickness. lab_ID is the code of the radiocarbon (C14) samples. C14_age is the radiocarbon (C14) dates. error is the error estimated in the radiocarbon (C14) datation of the samples. reservoir is to specify if the samples have marine reservoir effects. depth is the depth in cm of the radiocarbon samples. thickness is the tickness of the radiocarbon samples.
BACON format has 4 columns: labID, age, error, and depth. labID is the code of the radiocarbon (C14) samples. age is the radiocarbon (C14) date. error is the error estimated in the radiocarbon (C14) datation for the samples. depth is the depth in cm of the radiocarbon samples.

The function also produce the files required by CLAM or BACON in the working directory according to this folder structure: ~/{format}/Cores/{entity_number}/
The function creates a file .csv ({entity_number}.csv) with the radiocarbon data and two .txt ({entity_number}_depths.txt and {entity_number}_depthsID.txt). The last one is not used by CLAM or BACON but it might be useful to follow track of samples depths.
}
\description{
This function takes datation data, as returned by \code{\link[EPDr:getDatation]{getDatation}} and reshape it to comply with CLAM or BACON format.
}
\details{
The function has an interactive implementation. If some of the parameters \code{include_chron_not_in_c14}, \code{include_C14_not_in_chron}, \code{use_c14_conf_age}, and/or \code{use_c14_conf_depth} are not specified, the function will check the data and ask the user how it should proceed regarding some conflicts in the data between C14 data and chronologies.
}
\examples{
# TBW
}
\references{
\url{http://www.chrono.qub.ac.uk/blaauw/clam.html}

\url{http://chrono.qub.ac.uk/blaauw/bacon.html}
}

