% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr-export_functions.R
\name{exportDepths}
\alias{exportDepths}
\title{Reshape depths data to CLAM or BACON format}
\usage{
exportDepths(depths)
}
\arguments{
\item{depths}{Data frame with at least a column called depthcm.}
}
\value{
Vector with depths in ascending order.
}
\description{
This function takes depths data, as returned by \code{\link[EPDr:getPSamples]{getPSamples}}, to comply with CLAM or BACON format.
}
\examples{
#epd.connection <- connectToEPD(database="epd", user="epdr",
#                                 password="epdrpw", host="localhost")
#depths.1 <- getPSamples(1, epd.connection)
#exportDepths(depths.1)
#disconnectFromEPD(connection=epd.connection)
}
\references{
\url{http://www.chrono.qub.ac.uk/blaauw/clam.html}

\url{http://chrono.qub.ac.uk/blaauw/bacon.html}
}
