% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr-export_functions.R
\name{exportEvents}
\alias{exportEvents}
\title{Reshape events data to CLAM or BACON format}
\usage{
exportEvents(event, format = c("clam", "bacon"))
}
\arguments{
\item{event}{Data frame with event data for a particular core (entity) in the EPD DDBB.}

\item{format}{Character string indicating whether to export to "clam" or to "bacon" format.}
}
\value{
Data frame with event data in CLAM or BACON format.
}
\description{
This function takes event data to fit into a new table that comply with CLAM or BACON format.
}
\examples{
#epd.connection <- connectToEPD(host="localhost", database="epd_ddbb",
#                               user="epdr", password="epdrpw")
#event <- getEvents(51, epd.connection)
#event.clam <- exportEvents(event, format="clam")
#event.bacon <- exportEvents(event, format="bacon")

}
\references{
\url{http://www.chrono.qub.ac.uk/blaauw/clam.html}

\url{http://chrono.qub.ac.uk/blaauw/bacon.html}
}

