% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr-extract_functions.R
\docType{methods}
\name{extract_e}
\alias{extract_e}
\alias{extract_e,epd.entity-method}
\alias{extract_e-method}
\title{Extract Entity number from EPDr objects}
\usage{
extract_e(x)

\S4method{extract_e}{epd.entity}(x)
}
\arguments{
\item{x}{epd.entity Any object of classes \code{\link[EPDr:epd.entity]{epd.entity}} or
\code{\link[EPDr:epd.entity.df]{epd.entity.df}}.}
}
\value{
numeric. Value of the entity number.
}
\description{
This function looks into EPDr objects (i.e., \code{\link[EPDr]{epd.entity}}
or \code{\link[EPDr]{epd.entity.df}} classes), returning the entity 
number of the objects.
}
\examples{
##NOT TO RUN
# Not run
# epd.connection <- connectToEPD()
# epd.connection <- connectToEPD(database="epd", user="epdr",
#                                 password="epdrpw", host="localhost")
# epd.connection

# To list all the tables in the database we have connected with
# data <- getDatation(1, epd.connection)
# extract_e(data)
# [1] 1
# data <- getDatation(45, epd.connection)
# extract_e(data)
# [1] 45
# disconnectFromEPD(connection=epd.connection)

}
