% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr.R
\name{extractC14}
\alias{extractC14}
\title{Extract C14 data for a particular core (entity in the EPD DDBB)}
\usage{
extractC14(core_number, connection)
}
\arguments{
\item{core_number}{Integer or string with the core (entity) number for which C14 data want to be extracted}

\item{connection}{Connection object to a EPD DDBB where the query is made}
}
\value{
Matrix with
}
\description{
Given a core number (as in the EPD DDBB: e_) the function returns a matrix with the C14 data associated to this core
}
\examples{
# TBW
}

