% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr.R
\name{extractChronologies}
\alias{extractChronologies}
\title{Title}
\usage{
extractChronologies(core.num, conn)
}
\arguments{
\item{core.num}{TBW}

\item{conn}{TBW}
}
\value{
TBW
}
\description{
Title
}
\examples{
# TBW
}

