% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr.R
\name{extractEvents}
\alias{extractEvents}
\title{Extract events associated with a specific core (entity) in the EPD DDBB}
\usage{
extractEvents(core_number, connection)
}
\arguments{
\item{core_number}{Integer or string with the core (entity) number for which C14 data want to be extracted.}

\item{connection}{Connection object to a EPD DDBB where the query is made.}
}
\value{
NA or data frame, depending on whether there are events associated with the requested core. If there are events the
function return a data frame with combined information from synevent and event tables in the EPD. Columns in the data
frame follows terminology in the original database and are as follows:
\itemize{
 \item \code{event_}: Event identifier.
 \item \code{e_}: Core (entity) identifier.
 \item \code{depthcm}: Depth of the event in cm.
 \item \code{thickness}: Thickness of the event in cm.
 \item \code{event}: Code for the type of event. See EPD documentation for details.
 \item \code{name}: Name of the event.
 \item \code{agebp}: Known age of the event in BP.
 \item \code{ageuncertup}:
 \item \code{ageuncertlo}:
 \item \code{publ_}: Publication identifier.
}
}
\description{
Given a specific core number and connection to the EPD, this function return the information of events associated with the core.
}
\examples{
epd.connection <- connectToEPD(database="epd_ddbb", user="epdr",
                                 password="epdrpw", host="diegonl.ugr.es")
extractEvents(1, epd.connection)
extractEvents(51, epd.connection)
disconnectFromEPD(connection=epd.connection)
}

