% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr-standardize_methods.R
\docType{methods}
\name{filter_taxa}
\alias{filter_taxa}
\alias{filter_taxa,epd.entity.df,character,data.frame-method}
\alias{filter_taxa-method}
\alias{filter_taxa,epd.entity,character,data.frame-method}
\alias{filter_taxa-method}
\title{#' Expand EPDr objects with new taxa}
\usage{
filter_taxa(x, taxa, epd.taxonomy)

\S4method{filter_taxa}{epd.entity.df,character,data.frame}(x, taxa,
  epd.taxonomy)

\S4method{filter_taxa}{epd.entity,character,data.frame}(x, taxa, epd.taxonomy)
}
\arguments{
\item{x}{\code{\link[EPDr:epd.entity.df]{epd.entity.df}} or
\code{\link[EPDr:epd.entity]{epd.entity}} object to be modified.}

\item{taxa}{Character vector indicating the new taxa in the \code{@epd.entity.df} slot.}

\item{epd.taxonomy}{Data frame with the taxonomy from the EPD as from the
\code{\link[EPDr:get_taxonomy_epd]{get_taxonomy_epd}} function.}
}
\value{
\code{\link[EPDr:epd.entity.df]{epd.entity.df}} object with the modified
\code{@commdf@counts} slot.
}
\description{
This functions modifies \code{@commdf@counts} slot of
(\code{\link[EPDr:epd.entity.df]{epd.entity.df}} objects to filter taxa columns
to match \code{taxa}. The function add empty columns (\code{NA}) if a new taxa is
defined in \code{taxa} or remove columns for the taxa not included in \code{taxa}.
The function may look useless for a single entity but it is useful when standardizing
data from multiple entities.
}
\examples{
# epd.connection <- connectToEPD(host="localhost", database="epd",
# user="epdr", password="epdrpw")
# agedcounts.1 <- getAgedCounts(1, epd.connection)
# agedcounts.1@counts@taxa_names
# colnames(agedcounts.1@counts@counts)
# agedcounts.1.ft <- filter_taxa(agedcounts.1,
# c(agedcounts.1@counts@taxa_names, "prueba"), get_taxonomy_epd(epd.connection))
# colnames(agedcounts.1.ft@counts@counts)
# agedcounts.1.ft@counts@taxa_names
# head(agedcounts.1.ft@counts@counts)
}
