% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr-standardize_methods.R
\docType{methods}
\name{filter_taxa}
\alias{filter_taxa}
\alias{filter_taxa,epd.entity.df,character,data.frame-method}
\alias{filter_taxa,epd.entity,character,data.frame-method}
\title{Expand EPDr objects with new taxa}
\usage{
filter_taxa(x, taxa, epd.taxonomy, na_value = 0)

\S4method{filter_taxa}{epd.entity.df,character,data.frame}(x, taxa,
  epd.taxonomy, na_value = 0)

\S4method{filter_taxa}{epd.entity,character,data.frame}(x, taxa, epd.taxonomy,
  na_value = 0)
}
\arguments{
\item{x}{epd.entity \code{\link[EPDr]{epd.entity.df-class}} or
\code{\link[EPDr]{epd.entity-class}} object to be modified.}

\item{taxa}{character Character vector indicating the new 
taxa in the \code{@epd.entity.df} slot.}

\item{epd.taxonomy}{data.frame Data frame with the taxonomy from the 
EPD as from the \code{\link[EPDr]{get_taxonomy_epd}} function.}

\item{na_value}{numeric Number indicating the value to be used for
taxa not previously present in the entity.}
}
\value{
\code{\link[EPDr]{epd.entity.df-class}} object with the modified
\code{@commdf@counts} slot.
}
\description{
This function modifies \code{@commdf@counts} slot of
(\code{\link[EPDr]{epd.entity.df-class}} objects to filter taxa columns
to match \code{taxa}. The function adds empty columns (\code{NA}) if a new taxa is
defined in \code{taxa} or removes columns for the taxa not included in \code{taxa}.
The function may look useless for a single entity but it is useful when standardizing
data across multiple entities.
}
\examples{
\dontrun{
epd.connection <- connect_to_epd(host="localhost", database="epd",
                                 user="epdr", password="epdrpw")
epd.1 <- get_entity(1, epd.connection)
epd.1@commdf@taxanames
colnames(epd.1@commdf@counts)
epd.1.ft <- filter_taxa(epd.1, c(epd.1@commdf@taxa_names, "prueba"),
                        get_taxonomy_epd(epd.connection))
colnames(epd.1.ft@commdf@counts)
epd.1.ft@commdf@taxanames
head(epd.1.ft@commdf@counts)
}
}
