% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr-standardize_methods.R
\docType{methods}
\name{filter_taxagroups}
\alias{filter_taxagroups}
\alias{filter_taxagroups,epd.entity.df,character-method}
\alias{filter_taxagroups-method}
\alias{filter_taxagroups,epd.entity,character-method}
\alias{filter_taxagroups-method}
\title{Filter Taxa Groups}
\usage{
filter_taxagroups(x, taxagroups)

\S4method{filter_taxagroups}{epd.entity.df,character}(x, taxagroups)

\S4method{filter_taxagroups}{epd.entity,character}(x, taxagroups)
}
\arguments{
\item{x}{\code{\link[EPDr:epd.entity.df]{epd.entity.df}} or \code{\link[EPDr:epd.entity]{epd.entity}} object.}

\item{taxagroups}{Character vector indicating taxa groups to be selected}
}
\value{
The function returns a \code{\link[EPDr:epd.entity.df]{epd.entity.df}} with information in 
information on \code{@commdf@counts} slot only for taxa belonging to the specified taxa groups.
}
\description{
This function removes taxa from the slot @commdf@counts in the 
\code{\link[EPDr:epd.entity.df]{epd.entity.df}} object based on specified groups of taxa.
For example, the user can select to work only with pollen from trees and shrubs (TRSH)
or algae (ALGA). The function automatically remove counts from any particle added by 
the entity analyst to calculate particle concentrations in the samples 
(e.g., Lycopodium (added)). If a \code{\link[EPDr:epd.entity]{epd.entity}}, this is first converted 
into a \code{\link[EPDr:epd.entity.df]{epd.entity.df}} with \code{\link[EPDr:entity_to_matrices]{entity_to_matrices}}.
}
\details{
Details TBW
}
\examples{
# epd.connection <- connectToEPD(host="localhost", database="epd",
# user="epdr", password="epdrpw")
# agedcount <- getAgedCounts(1, epd.connection)
# agedcount.pollen <- filter_taxagroups(agedcount, "TRSH")
# str(agedcount@counts@counts)
# str(agedcount.pollen@counts@counts)
}
