% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr-standardize_methods.R
\docType{methods}
\name{filter_taxagroups}
\alias{filter_taxagroups}
\alias{filter_taxagroups,epd.entity.df,character-method}
\alias{filter_taxagroups,epd.entity,character-method}
\title{Filter Taxa Groups}
\usage{
filter_taxagroups(x, taxagroups)

\S4method{filter_taxagroups}{epd.entity.df,character}(x, taxagroups)

\S4method{filter_taxagroups}{epd.entity,character}(x, taxagroups)
}
\arguments{
\item{x}{epd.entity Object of class \code{\link[EPDr]{epd.entity.df}}
or \code{\link[EPDr]{epd.entity}} object.}

\item{taxagroups}{character Character vector indicating taxa groups 
to be selected.}
}
\value{
The function returns a \code{\link[EPDr]{epd.entity.df}} 
with information on \code{@commdf@counts} slot 
only for taxa belonging to the specified taxa groups.
}
\description{
This function removes taxa from the slot @commdf@counts in the 
\code{\link[EPDr]{epd.entity.df}} object based on specific groups of taxa.
For example, the user can select to work only with pollen from trees and 
shrubs (TRSH) or with algae (ALGA). If \code{x} is an 
\code{\link[EPDr]{epd.entity}} object, this is first converted into a 
\code{\link[EPDr]{epd.entity.df}} with 
\code{\link[EPDr]{entity_to_matrices}}.
}
\examples{
# epd.connection <- connect_to_epd(host="localhost", database="epd",
#                                  user="epdr", password="epdrpw")
# epd.1 <- get_entity(1, epd.connection)
# epd.1.trsh <- filter_taxagroups(epd.1, "TRSH")
# str(epd.1@commdf@counts)
# str(epd.1.trsh@commdf@counts)
}
