% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr-classes.R
\docType{class}
\name{geochron-class}
\alias{geochron-class}
\alias{geochron}
\title{Class for Geochronologies of an entity}
\description{
Class \code{geochron} stores all information relative to datation samples
and analysis in a specified entity of the European Pollen Database (EPD).
This object is created by \code{\link[EPDr]{get_geochron}}.
}
\section{Slots}{

\describe{
\item{\code{geochron}}{data.frame. Common information on geochron samples.}

\item{\code{aar}}{data.frame. AAR data for samples in geochron}

\item{\code{c14}}{data.frame. C14 data for samples in geochron}

\item{\code{esr}}{data.frame. ESR data for samples in geochron}

\item{\code{ft}}{data.frame. FT data for samples in geochron}

\item{\code{kar}}{data.frame. KAR data for samples in geochron}

\item{\code{pb210}}{data.frame. PB210 data for samples in geochron}

\item{\code{si32}}{data.frame. SI32 data for samples in geochron}

\item{\code{tl}}{data.frame. TL data for samples in geochron}

\item{\code{useries}}{data.frame. USERIES data for samples in geochron}

\item{\code{publ}}{data.frame PUBL data where the data were published}
}}

