% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr-get_functions.R
\name{get_publ}
\alias{get_publ}
\title{Retrieve publications from their publ ID number}
\usage{
get_publ(publ_, connection)
}
\arguments{
\item{publ_}{numeric with the publ_}

\item{connection}{PostgreSQLConnection. Object of class \code{PostgreSQLConnection} as returned by function \code{\link[EPDr:connect_to_epd]{connect_to_epd}}.}
}
\value{
data frame with information about the publication whole reference
}
\description{
This function is mainly intended for internal use. It retrieves information of publications by querying the database by the publication identification number.
}
\examples{
# Not run
# epd.connection <- connect_to_epd(database = "epd", user = "epdr",
#                                 password = "epdrpw", host = "localhost")
# get_publ(1, epd.connection)
# disconnectFromEPD(connection = epd.connection)
}
