% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr-get_functions.R
\name{get_samples}
\alias{get_samples}
\alias{.get_psamples}
\alias{.get_pagedpt}
\alias{.get_pcounts}
\alias{.get_pvars}
\alias{.get_pvtrans}
\alias{.get_syntype}
\alias{.get_pgroup}
\alias{.get_groups}
\title{Retrieve samples information for EPD entities}
\usage{
get_samples(e_, connection)

.get_psamples(e_, connection)

.get_pagedpt(e_, connection)

.get_pcounts(e_, connection)

.get_pvars(var_, connection)

.get_pvtrans(var_, connection)

.get_syntype(syntype, connection)

.get_pgroup(var_, connection)

.get_groups(groupid, connection)
}
\arguments{
\item{e_}{numeric Value indicating the entity number (e_) of the 
database that want to be queried.}

\item{connection}{PostgreSQLConnection Object of class 
\code{PostgreSQLConnection} as returned by function 
\code{\link[EPDr]{connect_to_epd}}.}

\item{var_}{numeric Value indicating the variable identification number
(var_) of the variable that is queried. Variables in the EPD are any
data that is counted and stored for each of the samples. They can be
biological particles (e.g., pollen), non-biological particles (e.g., 
added particles in the samples to calculate concentrations), or
summary of the previous computed afterwards (e.g., total pollen sum), 
among others.}

\item{syntype}{character One-character code indicating the type of 
synonym. It must be one of the following: M=morphological; 
N=nomenclatural; 1=name of a globally-monospecific genus (e.g. Zea); 
2=name of a genus that is monospecific within a continent, but not 
globally (e.g. Tilia americana in North America); or <null> for accepted 
variables.}

\item{groupid}{numeric Value indicating the group identification number
that is required.}
}
\value{
\code{\link[EPDr]{samples-class}} object. This is an EPDr object 
with information from different tables. See documentation of the EPD: 
\url{http://www.europeanpollendatabase.net/data/downloads/image/pollen-database-manual-20071011.doc}).
}
\description{
Functions in this group retrieve information relative to the samples
analized in an entity. The main function (\code{\link[EPDr]{get_samples}})
requires a valid connection to a database and the entity ID for the entity
of interest. All other functions (starting with a dot [.]) use different
arguments depending on the piece of information they retrieve
}
\section{get_samples}{

This function returns a \code{\link[EPDr]{samples-class}} object with several 
information from the rest of the functions in the group for a 
particular entity.
}

\section{.get_psamples}{

This functions returns a data.frame with information in in the P_SAMPLE table for the specified
entity. This corresponds with the sample depths for an entity. This table 
also stores an identifier for the worker who analyzed the sample.
}

\section{.get_pagedpt}{

This functions returns a data.frame with information in in the P_AGEDPT table for the specified
entity. This corresponds with fitted age estimates for the sample depths 
for each chronology specified for an entity. Names for each chronology 
are stored in the CHRON table. The age-depth data are generated by applying 
the age-model to the data in the AGEBASIS table.
}

\section{.get_pcounts}{

This functions returns a data.frame with information in in the P_COUNTS table for the specified
entity. This corresponds with the pollen counts for an entity. This table
stores the non-zero counts for each combination of variable and sample 
depth (the depths values are stored in table P_SAMPLE).
}

\section{.get_pvars}{

This functions returns a data.frame with information in in the P_VARS table for the specified
entity. This corresponds with details on the variables (taxa, charcoals, etc)
that have been counted in the biological samples (not for chronology).
}

\section{.get_pvtrans}{

This functions returns a data.frame with information in in the P_VTRANS table for the specified
entity. This corresponds with the names, codes and hierarchy information 
for each pollen variable recognized by the Pollen Database, 
including synonyms.
}

\section{.get_syntype}{

This functions returns a data.frame with information in in the SYNTYPE table for the specified
entity. This corresponds with full description of the synonym type codes 
used in P_VARS and M_VARS
}

\section{.get_pgroup}{

This functions returns a data.frame with information in in the P_GROUP table for the specified
entity. This corresponds with  group to which each pollen variable is 
assigned. These tables enable the user to define taxa into groups based 
on the life form and habitat of each taxon (e.g. trees, aquatics).
}

\section{.get_groups}{

This functions returns a data.frame with information in in the GROUPS table for the specified
entity. This corresponds with a look-up table for the names and codes for
variable groups.
}

\examples{
\dontrun{
epd.connection <- connect_to_epd(host = "localhost", database = "epd",
                                 user = "epdr", password = "epdrpw")
chron.400 <- get_samples(400, epd.connection)
chron.400
}
}
