% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr-get_functions.R
\name{get_workers}
\alias{get_workers}
\title{Title TBW}
\usage{
get_workers(worker_, connection)
}
\arguments{
\item{worker_}{TBW}

\item{connection}{TBW}
}
\value{
TBW
}
\description{
Title TBW
}
\examples{
# TBW
}
