% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr-get_functions.R
\name{getC14}
\alias{getC14}
\title{Extract C14 data for a particular core (entity in the EPD DDBB)}
\usage{
getC14(core_number, connection)
}
\arguments{
\item{core_number}{Integer with the core (entity) number for which C14 data want to be extracted.}

\item{connection}{Connection object to a EPD DDBB where the query is made.}
}
\value{
Data frame with combined information from c14 and geochron tables in the EPD. Columns in the data frame follows terminology
in the original database and are as follows:
\itemize{
 \item \code{e_}: Core (entity) identifier.
 \item \code{sample_}: C14 sample identifier.
 \item \code{agebp}: Uncalibrated C14 age.
 \item \code{agesdup}:
 \item \code{agesdlo}:
 \item \code{grthanage}:
 \item \code{basis}:
 \item \code{enriched}:
 \item \code{labnumber}:
 \item \code{deltac13}:
 \item \code{notes}:
 \item \code{method}:
 \item \code{depthcm}:
 \item \code{thickness}:
 \item \code{materialdated}:
 \item \code{publ_}: Publication identifier.
}
}
\description{
Given a core number (as in the EPD DDBB: e_) the function returns a matrix with the C14 data associated to this core. This values
come from two different tables of the EPD: c14 and geochron.
}
\examples{
#epd.connection <- connectToEPD(database="epd_ddbb", user="epdr",
#                                 password="epdrpw", host="localhost")
#getC14(1, epd.connection)
#getC14(400, epd.connection)
#disconnectFromEPD(connection=epd.connection)
}

