% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr-get_functions.R
\name{getChron}
\alias{getChron}
\alias{.getChron}
\alias{.getAgebound}
\alias{.getAgebasis}
\alias{.getRational}
\alias{.getAlsegs}
\alias{.getPAnldpt}
\alias{.getSynevent}
\alias{.getEvent}
\title{Retrieve chronological information for EPD entities}
\usage{
getChron(e_, connection)

.getChron(e_, connection)

.getAgebound(e_, connection)

.getAgebasis(e_, connection)

.getRational(rcode, connection)

.getAlsegs(e_, connection)

.getPAnldpt(e_, connection)

.getSynevent(e_, connection)

.getEvent(event_, connection)
}
\arguments{
\item{e_}{numeric. Value indicating the entity number (e_) of the database that want to be queried.}

\item{connection}{PostgreSQLConnection. Object of class \code{PostgreSQLConnection} as returned by function \code{\link[EPDr:connectToEPD]{connectToEPD}}.}

\item{rcode}{character. Three letter code for the rational.}

\item{event_}{numeric. Value indicating the event number (event_) of the database of the requested event.}
}
\value{
\code{\link[EPDr:chron]{chron}} object. This is an EPDr object with information from different tables. See documentation of the EPD: \url{http://www.europeanpollendatabase.net/data/downloads/image/pollen-database-manual-20071011.doc}).
}
\description{
Functions in this family retrieves information relative to the chronologies used to calculate samples ages for a particular entity. The main function (\code{\link[EPDr:getChron]{getChron}}) requires a valid connection to the database and the entity ID for the entity of interest. All other functions (starting with a dot [.]) use different arguments depending on the piece of information they retrieve.
}
\section{getChron}{

This function returns a \code{\link[EPDr:chron]{chron}} object with several information from the rest of the functions for a particular entity.
}

\section{.getChron}{

This function returns information in the CHRON table for the specified entity. This corresponds with chronologies for that entity.
}

\section{.getAgebound}{

This function returns information in the AGEBOUND table for the chronologies in that entity. This corresponds with age limits for that entity calculated according to each chronology.
}

\section{.getAgebasis}{

This function returns information in the AGEBASIS table for the specified entity. This corresponds with the depth and C14 data used to calibrate the age-depth model in each chronology.
}

\section{.getRational}{

This function returns information in the RATIONAL table for the specified entity. This corresponds with the rational to use each sample in the AGEBASIS table to calibrate the age-depth model.
}

\section{.getAlsegs}{

This function returns information in the ALSEGS table for the specified entity. This corresponds with segments of annual laminations in the entity.
}

\section{.getPAnldpt}{

This function returns information in the P_ANLDPT table for the specified entity. This corresponds with details on the lamination for each segment on the ALSEGS table for each entity.
}

\section{.getSynevent}{

This function returns information in the SYNEVENT table for the specified entity. This corresponds with the geological events that affect that entity.
}

\section{.getEvent}{

This function returns information in the EVENT table for the specified entity. This corresponds with details on the geological event specified for that entity in the SYNEVENT table.
}

\examples{
# Not run
# library(EPDr)
# epd.connection <- connectToEPD(host="localhost", database="epd",
#                               user="epdr", password="epdrpw")
# chron.400 <- getChron(400, epd.connection)
# chron.400
# disconnectFromEPD(epd.connection)
}
