% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr-get_functions.R
\name{getChronology}
\alias{getChronology}
\title{Extract chronologies associated with a core (entity) in the pollen database}
\usage{
getChronology(core_number, connection)
}
\arguments{
\item{core_number}{Integer with the core (entity) number for which C14 data want to be extracted.}

\item{connection}{Connection object to a EPD DDBB where the query is made.}
}
\value{
List with 5 elements:
\itemize{
  \item \code{number_of_chronologies}: Integer indicating the number of chronologies associated with the core (entity).
  \item \code{default_chronology}: Integer indicating which is the default chronology according to the EPD DDBB.
  \item \code{chron}: Data frame with the meta information on how each calibration was built.
  \item \code{agebasis}: Data frame with all the information (depth and age for C14 and no-C14 data) used to build the chronologies.
  \item \code{no_C14}: Data frame with no-C14 data used to build the chronologies.
}
}
\description{
Given a core (entity) number, \code{\link[EPDr:getChronology]{getChronology}} extract all the information about chronologies associated with this core
in the EPD DDBB. This information comes from two different tables in the DDBB: chron and agebasis.
}
\examples{
#epd.connection <- connectToEPD(database="epd_ddbb", user="epdr",
#                                 password="epdrpw", host="localhost")
#getChronology(1, epd.connection)
#getChronology(400, epd.connection)
#disconnectFromEPD(connection=epd.connection)
}

