% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr-get_functions.R
\name{getDepths}
\alias{getDepths}
\title{Depths of pollen samples}
\usage{
getDepths(e_, connection)
}
\arguments{
\item{e_}{numeric. Value indicating the entity number (e_) of the database that want to 
be queried.}

\item{connection}{PostgreSQLConnection. Object of class \code{PostgreSQLConnection} as returned by function \code{\link[EPDr:connectToEPD]{connectToEPD}}.}
}
\value{
Data frame with all the information for pollen samples as in the p_sample table of the EPD DDBB.
}
\description{
Given a specific core (entity) this function return the information at which depths samples were taken for pollen data.
}
\examples{
#epd.connection <- connectToEPD(database="epd", user="epdr",
#                                 password="epdrpw", host="localhost")
#getDepths(1, epd.connection)
#getDepths(51, epd.connection)
#disconnectFromEPD(connection=epd.connection)
}

