% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr-get_functions.R
\name{getGeochron}
\alias{.getAAR}
\alias{.getC14}
\alias{.getESR}
\alias{.getFT}
\alias{.getGeochron}
\alias{.getKAR}
\alias{.getPB210}
\alias{.getSI32}
\alias{.getTL}
\alias{.getUSERIES}
\alias{getGeochron}
\title{Retrieving information for an entity in the EPD}
\usage{
getGeochron(e_, connection)

.getGeochron(e_, connection)

.getAAR(e_, connection)

.getC14(e_, connection)

.getESR(e_, connection)

.getFT(e_, connection)

.getKAR(e_, connection)

.getPB210(e_, connection)

.getSI32(e_, connection)

.getTL(e_, connection)

.getUSERIES(e_, connection)
}
\arguments{
\item{e_}{numeric. Value indicating the entity number (e_) of the database that is queried.}

\item{connection}{PostgreSQLConnection. Object of class \code{PostgreSQLConnection} as returned by function \code{\link[EPDr:connectToEPD]{connectToEPD}}.}
}
\value{
Data frame with all information from specific tables in the EPD (see documentation of the EPD: \url{http://www.europeanpollendatabase.net/data/downloads/image/pollen-database-manual-20071011.doc}) for the requested entity. Columns names in the resulting data frames will vary among functions.
}
\description{
Functions in this group retrieve different sort of information from an specific entity in the database.
}
\details{
All functions here are designed to retrieve information from a single entity. If multiple entity numbers are requested, the functions return data only for the first one. Each function retrieve data from a specific table or set of tables that are conveniently combined if necessary.
}
\section{getGeochron}{

This function returns information in the GEOCHRON table for the specified entity. This corresponds with the common geochronological data for the entity that have been analysed for that particular entity.
}

\section{.getGeochron}{

This function returns information in the GEOCHRON table for the specified entity. This corresponds with the common geochronological data for the entity that have been analysed for that particular entity.
}

\section{.getAAR}{

This function returns information in the AAR table for the specified entity. This corresponds with Amino Acid Racemization data for datation samples.
}

\section{.getC14}{

This function returns information in the C14 table for the specified entity. This corresponds with C14 data for all radiocarbon samples that have been analysed for that particular entity.
}

\section{.getESR}{

This function returns information in the ESR table for the specified entity. This corresponds with Electron Spin Resonance data for datation samples.
}

\section{.getFT}{

This function returns information in the FT table for the specified entity. This corresponds with Fission Track data for datation samples.
}

\section{.getKAR}{

This function returns information in the KAR table for the specified entity. This corresponds with Fission Track data for datation samples.
}

\section{.getPB210}{

This function returns information in the PB210 table for the specified entity. This corresponds with \eqn{210^{Pb}}{[Pb]^210} Isotope data for datation samples.
}

\section{.getSI32}{

This function returns information in the SI32 table for the specified entity. This corresponds with Silicon-32 data for datation samples.
}

\section{.getTL}{

This function returns information in the TL table for the specified entity. This corresponds with Thermoluminescence data for datation samples.
}

\section{.getUSERIES}{

This function returns information in the USERIES table for the specified entity. This corresponds with Uranium-series data for datation samples.
}
\examples{
# Not run
# epd.connection <- connectToEPD(database="epd", user="epdr",
#                                 password="epdrpw", host="localhost")
# .getC14(1, epd.connection)
# .getC14(400, epd.connection)
#
# getGeochron(400, epd.connection)
# disconnectFromEPD(connection=epd.connection)


}

