% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr-get_functions.R
\name{getPSamples}
\alias{getPSamples}
\title{Query palynological samples of EPD entities}
\usage{
getPSamples(e_, connection)
}
\arguments{
\item{e_}{numeric. Value indicating the entity number (e_) of the database that want to 
be queried.}

\item{connection}{PostgreSQLConnection. Object of class \code{PostgreSQLConnection} as
returned by function \code{\link[EPDr:connectToEPD]{connectToEPD}}.}
}
\value{
data.frame. Data frame with all information from the P_SAMPLE table in the database
for that particular entity (see documentation of the EPD:
\url{http://www.europeanpollendatabase.net/data/downloads/image/pollen-database-manual-20071011.doc}).
}
\description{
This function queries the database to request all information about palynological samples 
of an specified entity. To perform the query the function requires the number
of the entity that want to be queried and a valid connection to the database. Hence,
the following parameters are mandatory:
}
\examples{
# Not run
# epd.connection <- connectToEPD(database="epd", user="epdr",
#                                 password="epdrpw", host="localhost")
# getPSamples(1, epd.connection)
# getPSamples(51, epd.connection)
# disconnectFromEPD(connection=epd.connection)

}
