% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr-get_functions.R
\name{getPubl}
\alias{getPubl}
\title{Retrieve publications from their publ ID number}
\usage{
getPubl(publ_, connection)
}
\arguments{
\item{publ_}{numeric with the publ_}

\item{connection}{PostgreSQLConnection. Object of class \code{PostgreSQLConnection} as returned by function \code{\link[EPDr:connectToEPD]{connectToEPD}}.}
}
\value{
data frame with information about the publication whole reference
}
\description{
This function is mainly intended for internal use. It retrieves information of publications by querying the database by the publication identification number.
}
\examples{
# Not run
# epd.connection <- connectToEPD(database="epd", user="epdr",
#                                 password="epdrpw", host="localhost")
# getPubl(1, epd.connection)
# disconnectFromEPD(connection=epd.connection)
}
