% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr-get_functions.R
\name{getRestriction}
\alias{getRestriction}
\title{Title TBW}
\usage{
getRestriction(e_, connection)
}
\arguments{
\item{e_}{TBW}

\item{connection}{TBW}
}
\value{
TBW
}
\description{
TBW
}
\examples{
# TBW
}

