% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr-get_functions.R
\name{getSite}
\alias{getSite}
\alias{.getSiteloc}
\alias{.getSitedesc}
\alias{.getSiteinfo}
\alias{.getPoldiv1}
\alias{.getPoldiv2}
\alias{.getPoldiv3}
\alias{.getIGCPtype}
\alias{.getInfotype}
\title{Retrieve site information for EPD entities}
\usage{
getSite(e_, connection)

.getSiteloc(site_, connection)

.getSitedesc(site_, connection)

.getSiteinfo(site_, connection)

.getPoldiv1(poldiv1_, connection)

.getPoldiv2(poldiv2_, poldiv1_, connection)

.getPoldiv3(poldiv3_, poldiv2_, poldiv1_, connection)

.getIGCPtype(igcptype, connection)

.getInfotype(icode, connection)
}
\arguments{
\item{e_}{numeric. Value indicating the entity number (e_) of the database that want to be queried.}

\item{connection}{PostgreSQLConnection. Object of class \code{PostgreSQLConnection} as returned by function \code{\link[EPDr:connectToEPD]{connectToEPD}}.}

\item{site_}{numeric. Value indicating the site number (site_) of interest in the database.}

\item{poldiv1_}{Three character string. The three character string are the international country code.}

\item{poldiv2_}{Two number string (character). This string with length equal two and with numbers represent the regions code for administrative regions in each country. The code is not unique so to capture an specific region in a country you need to provide always country code (poldiv1_) and region code (poldiv2_).}

\item{poldiv3_}{Three number string (character). This string with length equal three and with numbers represent the 3rd level administrative regions in each country.  The code is not unique so to capture an specific 3rd level region in a country you need to provide always country code (poldiv1_), region code (poldiv2_), and 3rd level region code (poldiv3_).}

\item{igcptype}{Character. Representing the IGCP type code.}

\item{icode}{Character. Three letter string representing the information code.}
}
\value{
\code{\link[EPDr:site]{site}} object. This is an EPDr object with information from different tables. See documentation of the EPD: \url{http://www.europeanpollendatabase.net/data/downloads/image/pollen-database-manual-20071011.doc}).
}
\description{
Functions in this family retrieves information relative to the site where the entities have been sampled. The main function (\code{\link[EPDr:getSite]{getSite}}) requires a valid connection to the database and the entity ID for the entity of interest. All other functions (starting with a dot [.]) use different arguments depending on the piece of information they retrieve.
}
\details{
Details
}
\section{getSite}{

This function returns a \code{\link[EPDr:site]{site}} object with several information from the rest of the functions for a particular entity.
}

\section{.getSiteloc}{

This function returns information in the SITELOC table for the specified entity. This corresponds with location data for the site wheres samples were taken.
}

\section{.getSitedesc}{

This function returns information in the SITEDESC table for the specified entity. This corresponds with a description of the site wheres samples were taken.
}

\section{.getSiteinfo}{

This function returns information in the SITEINFO table for the specified entity. This corresponds with a summary data of all type of information in the database for that particular entity (chronological, palynological, etc).
}

\section{.getPoldiv1}{

This function returns information in the POLDIV1 table for the specified country (poldiv1_ is the country code). This corresponds with data of the country in which a site belong to.
}

\section{.getPoldiv2}{

This function returns information in the POLDIV2 table for the specified region (poldiv2_ is the region code). This corresponds with data of the region in which a site belong to.
}

\section{.getPoldiv3}{

This function returns information in the POLDIV3 table for the specified 3rd level region (poldiv3_ is the 3rd level region code). This corresponds with data of the 3rd level region in which a site belong to.
}

\section{.getIGCPtype}{

This function returns information in the IGCPTYPE table for the specified IGCP region. This corresponds with data of the IGCP region in which a site belong to.
}

\section{.getInfotype}{

This function returns information in the INFOTYPE table for the specified info code (icode). This corresponds with a longer description of the info type codes returned by \code{.getSiteinfo}.
}

\examples{
# Not run
# library(EPDr)
# epd.connection <- connectToEPD(host="localhost", database="epd",
#                               user="epdr", password="epdrpw")
# site.400 <- getSite(400, epd.connection)
# site.400
# disconnectFromEPD(epd.connection)
}
