% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr-get_functions.R
\name{getSite_old}
\alias{getSite_old}
\title{Query Site information of EPD entities}
\usage{
getSite_old(e_, connection)
}
\arguments{
\item{e_}{numeric. Value indicating the entity number (e_) of the database that want to 
be queried.}

\item{connection}{PostgreSQLConnection. Object of class \code{PostgreSQLConnection} as returned
by function \code{\link[EPDr:connectToEPD]{connectToEPD}}.}
}
\value{
data.frame The function return a data.frame with all the information on the SITELOC
table of the database (see documentation of the EPD:
\url{http://www.europeanpollendatabase.net/data/downloads/image/pollen-database-manual-20071011.doc}).
}
\description{
This function allows to query the database to request all information about the site in
which an entity was sampled. To perform the query the function requires the number
of the entity that want to be queried and a valid connection to the database. Hence,
the following parameters are mandatory:
}
\examples{
# Not run
# library(EPDr)
# epd.connection <- connectToEPD(host="localhost", database="epd",
#                               user="epdr", password="epdrpw")
# site.400 <- getSite(400, epd.connection)
# site.400
# disconnectFromEPD(epd.connection)

}
