% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr.R
\name{getSiteForClam}
\alias{getSiteForClam}
\title{Title}
\usage{
getSiteForClam(site.num, ddbb, get.dephts = TRUE)
}
\arguments{
\item{site.num}{TBC}

\item{ddbb}{TBC}

\item{get.dephts}{TBC}
}
\value{
TBC
}
\description{
Title
}
\examples{
# TBC
}

