% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr-get_functions.R
\name{getTaxonomyEPD}
\alias{getTaxonomyEPD}
\title{Query the taxonomy table of the EPD}
\usage{
getTaxonomyEPD(connection)
}
\arguments{
\item{connection}{PostgreSQLConnection. Object of class \code{PostgreSQLConnection} as returned by function \code{\link[EPDr:connectToEPD]{connectToEPD}}.}
}
\value{
data.frame The function return a data.frame with the combined information from P_VARS and P_GROUP tables of the database (see documentation of the EPD: \url{http://www.europeanpollendatabase.net/data/downloads/image/pollen-database-manual-20071011.doc}).
}
\description{
The function queries the whole taxonomy of the database by combining 
information from the P_VARS and P_GROUP tables. Because it queries all 
the records in the database it only requires a valid connection to the 
database server as parameter.
}
\details{
Some users may find this function useful but it has been mainly 
implemented to be used by other functions in the \code{EPDr} package.
}
\examples{
# Not run
# library(EPDr)
# epd.connection <- connectToEPD(host="localhost", database="epd_ddbb",
#                               user="epdr", password="epdrpw")
# epd.taxonomy <- getTaxonomy(epd.connection)
# epd.taxonomy
# disconnectFromEPD(epd.connection)

}

