% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr-standardize_methods.R
\docType{methods}
\name{giesecke_default_chron}
\alias{giesecke_default_chron}
\alias{giesecke_default_chron,epd.entity-method}
\alias{giesecke_default_chron-method}
\title{Make Giesecke the default chronology}
\usage{
giesecke_default_chron(x)

\S4method{giesecke_default_chron}{epd.entity}(x)
}
\arguments{
\item{x}{\code{\link[EPDr:epd.entity]{epd.entity}} or \code{\link[EPDr:epd.entity.df]{epd.entity.df}} objects.}
}
\value{
The function returns a modified version of the \code{epd.entity} in which the
slot @default_chronology is changed to \code{9999} if they recalculated ages for this
entity. A numeric code is used (\code{9999}) because of the object definition is
numeric. This values is used in other functions to grab the right ages from Giesecke.
}
\description{
This function makes chronologies from Giesecke the default to be used in
\code{\link[EPDr:epd.entity]{epd.entity}} and \code{\link[EPDr:epd.entity.df]{epd.entity.df}}
objects if they are available for that particular entity.
}
\details{
Details TBW
}
\examples{
# epd.connection <- connectToEPD(host="localhost", database="epd",
# user="epdr", password="epdrpw")
# agedcount <- getAgedCounts(1, epd.connection)
# agedcount@ages@default_chronology
# agedcount <- giesecke_default_chron(agedcount)
# agedcount@ages@default_chronology
}
