% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr-standardize_methods.R
\docType{methods}
\name{giesecke_default_chron}
\alias{giesecke_default_chron}
\alias{giesecke_default_chron,epd.entity-method}
\title{Make Giesecke the default chronology}
\usage{
giesecke_default_chron(x)

\S4method{giesecke_default_chron}{epd.entity}(x)
}
\arguments{
\item{x}{epd.entity \code{\link[EPDr]{epd.entity-class}} or 
\code{\link[EPDr]{epd.entity.df-class}} objects.}
}
\value{
The function returns a modified version of the 
\code{epd.entity} in which the slot @defaultchron is changed to 
\code{9999} if they recalculated ages for this entity. A numeric code 
is used (\code{9999}) because of the object definition is numeric. This 
values is used in other functions to grab the right ages from Giesecke.
}
\description{
This function makes chronologies from Giesecke the default to be used in
\code{\link[EPDr]{epd.entity-class}} and \code{\link[EPDr]{epd.entity.df-class}}
objects if they are available for that particular entity.
}
\details{
Details The function first check whether that entity has chronological
data in Giesecke et al. (2013), and then, if available, change the
default chronology number from the EPD to the one provided by Giesecke
(i.e., x@defaultchron == 9999)
}
\examples{
\dontrun{
epd.connection <- connect_to_epd(host="localhost", database="epd",
                                 user="epdr", password="epdrpw")
epd.1 <- get_entity(1, epd.connection)
epd.1@defaultchron
epd.1 <- giesecke_default_chron(epd.1)
epd.1@defaultchron
}
}
\references{
Giesecke, T., Davis, B., Brewer, S., Finsinger, W., 
Wolters, S., Blaaw, M., de Beaulieu, J.L., Binney, H., Fyfe, R.M.,
Gaillard, M.J., Gil-Romera, G., van der Knaap, W.O. Kunes, P.,
Kuhl, N., van Leeuwen, J.F.N, Leydet, M., Lotter, A.F., Ortu, E.,
Semmler, M., and Bradshaw, R.H.W (2013). Towards mapping the late
Quaternary vegetation change of Europe. Vegetation History and
Archaeobotany, 23, 75-86.
}
