% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr-standardize_methods.R
\docType{methods}
\name{intervals_counts}
\alias{intervals_counts}
\alias{intervals_counts,epd.entity.df,numeric,numeric-method}
\alias{intervals_counts,epd.entity,numeric,numeric-method}
\title{Mean counts for specific time intervals}
\usage{
intervals_counts(x, tmin, tmax, ..., newlabels = NULL, chronology = NULL)

\S4method{intervals_counts}{epd.entity.df,numeric,numeric}(x, tmin, tmax, ...,
  newlabels = NULL, chronology = NULL)

\S4method{intervals_counts}{epd.entity,numeric,numeric}(x, tmin, tmax, ...,
  newlabels = NULL, chronology = NULL)
}
\arguments{
\item{x}{epd.entity.df An \code{\link[EPDr]{epd.entity.df}} object as 
returned by the \code{\link[EPDr]{entity_to_matrices}} function.}

\item{tmin}{numeric Numeric vector indicating the lower limits 
(in years cal. BP) for the time intervals.}

\item{tmax}{numeric Numeric vector indicating the upper limits 
(in years cal. BP) for the time intervals}

\item{...}{Not used with current methods.}

\item{newlabels}{character Character vector with labels for each time 
intervals, if none are provided the functions generate them with the 
following format \code{tmin}-\code{tmax}.}

\item{chronology}{numeric Number specifying the chronology from which 
ages should be used to calculate the interpolations. If none is provided 
the function uses the default chronology from the object (see
\code{\link[EPDr]{giesecke_default_chron}}).}
}
\value{
The function returns a \code{\link[EPDr]{epd.entity.df}} object, 
similar to \code{epd.entity.df} in which ages and counts has been modified
to the time intervarls specified and the counts estimated for these periods.
}
\description{
This function uses data (sample ages and sample counts) from an
\code{\link[EPDr]{epd.entity.df}} object to calculate mean counts
for samples within specific time intervals defined by the user. This can be
used to estimate mean counts for the same time intervals for multiple 
entities or cores in the database, standardizing them for 
integrative analysis.
}
\details{
Time intervals without sample (data) in the entity are fill with 
\code{NA}. This is convenient if analysis are carried out with 
multiple entities.
}
\examples{
\dontrun{
epd.connection <- connect_to_epd(host="localhost", database="epd",
                                 user="epdr", password="epdrpw")
epd.1 <- get_entity(1, epd.connection)
epd.1.int <- intervals_counts(epd.1, tmin = seq(0, 21000, by = 1000),
                              tmax = seq(999, 21999, by = 1000))
epd.3 <- get_entity(3, epd.connection)
epd.3.int <- intervals_counts(epd.3, tmin = seq(0, 21000, by = 1000),
                              tmax = seq(999, 21999, by = 1000))
epd.3.int <- intervals_counts(epd.3, tmin = seq(0, 21000, by = 1000),
                              tmax = seq(999, 21999, by = 1000), 2)
}
}
