% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr-standardize_functions.R
\name{intervalsCounts}
\alias{intervalsCounts}
\title{Title TBW}
\usage{
intervalsCounts(agedcount, tmin, tmax, labels = NULL, chronology = NULL)
}
\arguments{
\item{agedcount}{TBW}

\item{tmin}{TBW}

\item{tmax}{TBW}

\item{labels}{TBW}

\item{chronology}{TBW}
}
\value{
TBW
}
\description{
Title TBW
}
\examples{
# TBW
}

