% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr-list_functions.R
\name{list_e}
\alias{list_e}
\title{Title}
\usage{
list_e(connection, countries = NULL, regions = NULL, sites = NULL,
  coords = NULL)
}
\arguments{
\item{connection}{}

\item{countries}{}

\item{regions}{}

\item{sites}{}

\item{coords}{}
}
\description{
Title
}

