% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr-list_functions.R
\name{list_publ}
\alias{list_publ}
\title{Title TBW}
\usage{
list_publ(connection, e_ = NULL, countries = NULL, regions = NULL,
  sites = NULL, coords = NULL)
}
\arguments{
\item{connection}{TBW}

\item{e_}{TBW}

\item{countries}{TBW}

\item{regions}{TBW}

\item{sites}{TBW}

\item{coords}{TBW}
}
\value{
TBW
}
\description{
Title TBW
}
\examples{
# TBW
}

