% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr-list_functions.R
\name{list_sites}
\alias{list_sites}
\title{Title TBW}
\usage{
list_sites(connection, countries = NULL, regions = NULL, coords = NULL)
}
\arguments{
\item{connection}{TBW}

\item{countries}{TBW}

\item{regions}{TBW}

\item{coords}{TBW}
}
\value{
TBW
}
\description{
Title TBW
}
\examples{
# TBW
}

