% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr-list_functions.R
\name{list_taxa}
\alias{list_taxa}
\title{Title TBW}
\usage{
list_taxa(connection, groups = NULL)
}
\arguments{
\item{connection}{TBW}

\item{groups}{TBW}
}
\value{
TBW
}
\description{
Title TBW
}
\examples{
# TBW
}

