% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr-list_functions.R
\name{list_taxagroups}
\alias{list_taxagroups}
\title{Title TBW}
\usage{
list_taxagroups(connection)
}
\arguments{
\item{connection}{TBW}
}
\value{
TBW
}
\description{
Title TBW
}
\examples{
# TBW
}

