% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr-plotting_functions.R
\name{map_by_TaxaAge}
\alias{map_by_TaxaAge}
\title{Title TBW}
\usage{
map_by_TaxaAge(agedcounts, taxa, sample_label, pres_abse = FALSE,
  pollen_thres = NULL, zoom_coord = NULL, points_pch = 21,
  points_colour = NULL, points_fill = NULL, points_range_size = NULL,
  map_title = NULL, legend_range = NULL, legend_title = NULL,
  napoints_size = 0.75, napoints_pch = 19, napoints_colour = "grey45",
  napoints_fill = NA, countries_bg_colour = "grey80",
  countries_border_colour = "grey90")
}
\arguments{
\item{agedcounts}{TBW}

\item{taxa}{TBW}

\item{sample_label}{TBW}

\item{pres_abse}{TBW}

\item{pollen_thres}{TBW}

\item{zoom_coord}{TBW}

\item{points_pch}{TBW}

\item{points_colour}{TBW}

\item{points_fill}{TBW}

\item{points_range_size}{TBW}

\item{map_title}{TBW}

\item{legend_range}{TBW}

\item{legend_title}{TBW}

\item{napoints_size}{TBW}

\item{napoints_pch}{TBW}

\item{napoints_colour}{TBW}

\item{napoints_fill}{TBW}

\item{countries_bg_colour}{TBW}

\item{countries_border_colour}{TBW}
}
\value{
TBW
}
\description{
Title TBW
}
\examples{
# TBW
}

