% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clam.R
\name{mix.curves}
\alias{mix.curves}
\title{Title TBW}
\usage{
mix.curves(ratio = 0.5, cc1 = "IntCal13.14C", cc2 = "Marine13.14C",
  offset = c(0, 0))
}
\arguments{
\item{ratio}{TBW}

\item{cc1}{TBW}

\item{cc2}{TBW}

\item{offset}{TBW}
}
\value{
TBW
}
\description{
Title TBW
}
\examples{
#TBW
}

