% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr.R
\name{noC14toCLAM}
\alias{noC14toCLAM}
\title{Reshape no-C14 data to CLAM format}
\usage{
noC14toCLAM(noC14)
}
\arguments{
\item{noC14}{Data frame with no-C14 data as those in a chronology list returned by \code{\link[EPDr:getChronologies]{getChronologies}}.}
}
\value{
Data frame with no-C14 data in CLAM format. This data frame can be easily combined with C14 data from \code{\link[EPDr:getChronologies]{C14toCLAM}} using \code{rbind}.
}
\description{
This function takes no-C14 data, that can be extracted from a chronology object, to fit into a new table that comply with
CLAM format.
}
\examples{
library(EPDr)
epd.connection <- connectToEPD(host="diegonl.ugr.es", database="epd_ddbb",
                               user="epdr", password="epdrpw")
c14 <- getC14(400, epd.connection)
c14.clam <- C14toCLAM(c14)
chron <- getChronologies(400, epd.connection)
noc14.clam <- noC14toCLAM(chron$no_C14)
all.clam <- rbind(c14.clam, noc14.clam)
}

