% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clam.R
\name{plot.proxies}
\alias{plot.proxies}
\title{Title #TBW}
\usage{
plot.proxies(prox, errors = TRUE, proxcol = grey(0.5), revyr = TRUE)
}
\arguments{
\item{prox}{#TBW}

\item{errors}{#TBW}

\item{proxcol}{#TBW}

\item{revyr}{#TBW}
}
\value{
#TBW
}
\description{
Title #TBW
}
\examples{
 #TBW
}

