% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr-plotting_functions.R
\docType{methods}
\name{plot_diagram}
\alias{plot_diagram}
\alias{plot_diagram,epd.entity.df-method}
\title{Plot pollen diagram of an entity}
\usage{
plot_diagram(x, ..., chronology = NULL, use_ages = TRUE,
  low_values_exag = TRUE, color_by_group = TRUE, order_taxa = TRUE)

\S4method{plot_diagram}{epd.entity.df}(x, ..., chronology = NULL,
  use_ages = TRUE, low_values_exag = TRUE, color_by_group = TRUE,
  order_taxa = TRUE)
}
\arguments{
\item{x}{TBW}

\item{...}{TBW}

\item{chronology}{TBW}

\item{use_ages}{TBW}

\item{low_values_exag}{TBW}

\item{color_by_group}{TBW}

\item{order_taxa}{TBW}
}
\value{
TBW
}
\description{
Description TBW
}
\details{
Details TBW
}
\examples{
# TBW
}
\references{
\url{http://blarquez.com/684-2/}
}
