% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clam.R
\name{pMC.age}
\alias{pMC.age}
\title{Title #TBW}
\usage{
pMC.age(mn, sdev, ratio = 100, decimals = 0)
}
\arguments{
\item{mn}{#TBW}

\item{sdev}{#TBW}

\item{ratio}{#TBW}

\item{decimals}{#TBW}
}
\value{
#TBW
}
\description{
Title #TBW
}
\examples{
 #TBW
}

