% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr-standardize_functions.R
\name{qualityIndex}
\alias{qualityIndex}
\title{Title TBW}
\usage{
qualityIndex(agedcount, datation, max_sample_dist = 2000,
  max_c14_dist = 5000)
}
\arguments{
\item{agedcount}{TBW}

\item{datation}{TBW}

\item{max_sample_dist}{TBW}

\item{max_c14_dist}{TBW}
}
\value{
TBW
}
\description{
From the Ecography 2013 paper, Appendix 3: For each site at a particular 1 kyr
time period, site data-quality was calculated as the mean normalized distance
of the nearest pollen sample and the nearest chronological control.  We
calculated the distance in years of the nearest pollen sample and the nearest
chronological control to each 1 kyr time period.  We eliminated sites where
the nearest pollen sample was over 2000 years away or the nearest
chronological control was over 5000 years away.  For the remaining sites in
each 1 kyr period, we created a summary measure of site data-quality by 
rescaling the two distances in years to a 0 - 1 scale and calculating the
mean.  For example, if the nearest sample to the 1 kyr BP time period at a
given site was at 1.050 kyr BP and the nearest chronological control was at
1.100 kyr BP, the raw distances would be 50 years and 100 years, respectively.
These equate to scaled values of 0.975 (i.e., 1 - 50/2000) and 0.98 (i.e., 1 -
100/5000) for sample and chronological quality, respectively, with a mean
data-quality for this site at the 1 kyr BP time period of 0.9775.
}
\examples{
# TBW
}

