% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr-manipulate_lists_functions.R
\name{remove_restricted}
\alias{remove_restricted}
\title{Remove restricted data from a list of objects}
\usage{
remove_restricted(list)
}
\arguments{
\item{list}{List of EPDr objects (\code{\link[EPDr]{epd.entity.df-class}}, or
\code{\link[EPDr]{epd.entity-class}}).}
}
\value{
List of EPDr objects.
}
\description{
This function is designed to work with list of EPDr objects
(\code{\link[EPDr]{epd.entity.df-class}}). The function parses all the elements
on the list and remove those with restriction on their use. See
\code{\link[EPDr]{check_restriction}}.
}
\examples{
\dontrun{
epd.connection <- connect_to_epd(host="localhost", database="epd",
                                 user="epdr", password="epdrpw")
e.list <- list_e(epd.connection, country=c("Spain"))
e.list <- e.list$e_
epd.spain <- lapply(e.list, get_entity, epd.connection)
epd.spain.un <- remove_restricted(epd.spain)
length(epd.spain)
length(epd.spain.un)
}
}
