% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr-manipulate_lists_functions.R
\name{remove_wo_counts}
\alias{remove_wo_counts}
\title{Remove data without counts from a list of objects}
\usage{
remove_wo_counts(list)
}
\arguments{
\item{list}{List of \code{\link[EPDr]{epd.entity.df-class}} objects.}
}
\value{
List of \code{\link[EPDr]{epd.entity.df-class}} objects, which is a 
subset of 'list' with counts data.
}
\description{
This function is designed to work with list of
\code{\link[EPDr]{epd.entity.df-class}} objects.
The function parses all the elements on the list and remove those without 
counts data for any taxa.
}
\examples{
\dontrun{
epd.connection <- connect_to_epd(host="localhost", database="epd",
                               user="epdr", password="epdrpw")
e.list <- listE(epd.connection, country=c("France"))
e.list <- e.list$e_
france <- lapply(e.list, get_entity, epd.connection)
france.wo <- remove_wo_counts(france)
length(france)
length(france.wo)
}
}
