% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr-standardize_functions.R
\name{removeWithoutAges}
\alias{removeWithoutAges}
\title{Remove data without ages from a list of objects}
\usage{
removeWithoutAges(objects)
}
\arguments{
\item{objects}{List of EPDr objects (\code{\link[EPDr:ages]{ages}},
\code{\link[EPDr:agedcounts]{agedcounts}}, \code{\link[EPDr:chronology]{chronology}},
and \code{\link[EPDr:datation]{datation}})}
}
\value{
List of EPDr objects.
}
\description{
This function is designed to work with list of EPDr objects
(\code{\link[EPDr:ages]{ages}}, \code{\link[EPDr:agedcounts]{agedcounts}},
\code{\link[EPDr:chronology]{chronology}}, and \code{\link[EPDr:datation]{datation}}),
but not (\code{\link[EPDr:counts]{counts}}). The function parses
all the elements on the list and remove those without a default chronology.
}
\examples{
# epd.connection <- connectToEPD(host="localhost", database="epd",
#                                user="epdr", password="epdrpw")
# e.list <- listE(epd.connection, country=c("Spain"))
# e.list <- e.list$e_
# spain.agedcounts <- lapply(e.list, getAgedCounts, epd.connection)
# spain.agedcounts.wo <- removeWithoutAges(spain.agedcounts)
# length(spain.agedcounts)
# length(spain.agedcounts.wo)

}

