% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr-classes.R
\docType{class}
\name{samples-class}
\alias{samples-class}
\alias{samples}
\title{Class for biological samples of an entity}
\description{
Class "samples" store all information about biological samples relative
to a specified entity of the European Pollen Database (EPD). This object
is created by \code{\link[EPDr]{get_samples}}.
}
\section{Slots}{

\describe{
\item{\code{psamples}}{data.frame. Information describing the samples: their 
depth (cm), thickness (cm), analyst, date of analysis, and notes.}

\item{\code{analysts}}{data.frame. Information on the worker that analyzed the 
data.}

\item{\code{pagedpt}}{data.frame. Estimated ages for each sample according to 
each available chronology in the database.}

\item{\code{pcounts}}{data.frame. Raw counts for each sample and each observed 
variable (taxon).}

\item{\code{pvars}}{data.frame. Data on the identification of each variable 
(taxon) in \code{@pcounts}.}

\item{\code{syntype}}{data.frame. Synonym type for those variables that are 
marked as synonyms.}

\item{\code{pvtrans}}{data.frame. Alternative names for each variable. Not 
used in the EPD.}

\item{\code{pgroup}}{data.frame. Linking table to identify each taxon (or 
variable) in a specific group.}

\item{\code{groups}}{data.frame. Details and longer description of the 
groups identified in \code{@pgroup}.}
}}

