% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr-classes.R
\docType{class}
\name{samplesdf-class}
\alias{samplesdf-class}
\alias{samplesdf}
\title{Class for samplesdf objects}
\description{
This object stores information regarding the biological samples in an
entity but in vector format rather than the original format in the 
EPD database. Note, however, that data here comes from the table
\code{psamples} in the database or an \code{\link[EPDr]{epd.entity}} 
object. Contrary to data in \code{\link[EPDr]{epd.entity}} objects, 
this are intended to be manipulated and modified if necessary.
}
\section{Slots}{

\describe{
\item{\code{sample_}}{numeric. Identification numbers for each sample.}

\item{\code{samplelabel}}{character. Name of each sample. Most time
they are a character version of the \code{@sample_} slot, but some others
are composites created by other functions in the \code{EPDr} package.}
}}

