% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clam.R
\name{student.t}
\alias{student.t}
\title{Title #TBW}
\usage{
student.t(y = 2450, error = 50, t.a = 3, t.b = 4, cc = 1,
  postbomb = c(), cc1 = "IntCal13", cc2 = "Marine13", cc3 = "SHCal13",
  cc4 = "ConstCal", Cutoff = 1e-05)
}
\arguments{
\item{y}{#TBW}

\item{error}{#TBW}

\item{t.a}{#TBW}

\item{t.b}{#TBW}

\item{cc}{#TBW}

\item{postbomb}{#TBW}

\item{cc1}{#TBW}

\item{cc2}{#TBW}

\item{cc3}{#TBW}

\item{cc4}{#TBW}

\item{Cutoff}{#TBW}
}
\value{
#TBW
}
\description{
Title #TBW
}
\examples{
 #TBW
}

