% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr-standardize_methods.R
\docType{methods}
\name{table_by_taxa_age}
\alias{table_by_taxa_age}
\alias{table_by_taxa_age,epd.entity.df,character,character-method}
\alias{table_by_taxa_age-method}
\alias{table_by_taxa_age,epd.entity,character,character-method}
\alias{table_by_taxa_age-method}
\title{Tabulate counts by taxa and age}
\usage{
table_by_taxa_age(x, taxa, sample_label)

\S4method{table_by_taxa_age}{epd.entity.df,character,character}(x, taxa,
  sample_label)

\S4method{table_by_taxa_age}{epd.entity,character,character}(x, taxa,
  sample_label)
}
\arguments{
\item{x}{\code{\link[EPDr:epd.entity.df]{epd.entity.df}} object where to extract the
data from, or a \code{\link[EPDr:epd.entity]{epd.entity}} that is automatically
transformed into a \code{\link[EPDr:epd.entity.df]{epd.entity.df}} and then tabulated.}

\item{taxa}{Character vector indicating the taxa to be included in the table. Several
taxa can be specified but the function returns data for only one taxa by summing all
counts, and the taxa name specified in the output is the first one in \code{taxa}.}

\item{sample_label}{Character vector indicating the ages or time intervals to be 
included in the table.}
}
\value{
Data frame with five columns:
\itemize{
\item{"e_"}{Entity identification number.}
\item{"londd"}{Longitude of the site in decimal degrees.}
\item{"latdd"}{Latitude of the site in decimal degrees.}
\item{"count"}{The count of that taxon in that particular sample (age or time interval).}
\item{"sample_label"}{The sample (age or time interval) at which the particulates
were counted.}
\item{"taxa_label"}{The taxa that has been counted.}
}
}
\description{
This function tabulate data from the \code{@commdf@counts} slot in EPDr objects
(\code{\link[EPDr:epd.entity.df]{epd.entity.df}}, or calculate them from a
\code{\link[EPDr:epd.entity]{epd.entity}} to summarize information for particular
taxa at particular age or time intervals (samples). This function is useful to reshape
the data to be plotted and mapped by \code{link[ggplot]{ggplot}}. It was written to be
used by \code{\link[EPDr:map_taxa_age]{map_taxa_age}} function.
}
\examples{
# TBW
}
