% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr-standardize_methods.R
\docType{methods}
\name{table_by_taxa_age}
\alias{table_by_taxa_age}
\alias{table_by_taxa_age,epd.entity.df,character,character-method}
\alias{table_by_taxa_age,epd.entity,character,character-method}
\title{Tabulate counts by taxa and age}
\usage{
table_by_taxa_age(x, taxa, sample_label)

\S4method{table_by_taxa_age}{epd.entity.df,character,character}(x, taxa,
  sample_label)

\S4method{table_by_taxa_age}{epd.entity,character,character}(x, taxa,
  sample_label)
}
\arguments{
\item{x}{epd.entity.df \code{\link[EPDr]{epd.entity.df}} object where 
to extract the data from, or a \code{\link[EPDr]{epd.entity}} that is 
automatically transformed into a \code{\link[EPDr]{epd.entity.df}} and 
then tabulated.}

\item{taxa}{character Character vector indicating the taxa to be included 
in the table. Several taxa can be specified but the function returns data 
for only one taxa by summing all counts, and the taxa name specified 
in the output is the first one in \code{taxa}. This is useful when
you want to combine data from different taxa in the same genus,
for instance.}

\item{sample_label}{character Character vector indicating the ages or time 
intervals to be included in the table.}
}
\value{
Data frame with six columns: \code{e_}, \code{londd}, \code{latdd},
\code{count}, \code{sample_label}, and \code{taxa_label}.
\itemize{
\item{"e_"}{Entity identification number.}
\item{"londd"}{Longitude of the site in decimal degrees.}
\item{"latdd"}{Latitude of the site in decimal degrees.}
\item{"count"}{The count of that taxon in that particular sample
(age or time interval).}
\item{"sample_label"}{The sample (age or time interval) at which the
particulates were counted.}
\item{"taxa_label"}{The taxa that has been counted.}
}
}
\description{
This function tabulates data from the \code{@commdf@counts} slot 
in EPDr objects (\code{\link[EPDr]{epd.entity.df}} or calculated 
them from a \code{\link[EPDr]{epd.entity}}) to summarize information 
for particular taxa at particular age or time intervals (samples). 
This function is useful to reshape the data to be plotted and mapped 
by \code{link[ggplot2]{ggplot}}. It was written to be used by 
\code{\link[EPDr]{map_taxa_age}} function.
}
\examples{
\dontrun{
epd.connection <- connect_to_epd(host = "localhost", database = "epd",
                               user = "epdr", password = "epdrpw")
epd.1 <- get_entity(1, epd.connection)
epd.1 <- filter_taxagroups(epd.1, c("HERB", "TRSH", "DWAR",
                                      "LIAN", "HEMI", "UPHE"))
epd.1 <- giesecke_default_chron(epd.1)

epd.1 <- interpolate_counts(epd.1, seq(0, 22000, by = 1000))
epd.taxonomy <- getTaxonomyEPD(epd.connection)
epd.1 <- taxa_to_acceptedtaxa(epd.1, epd.taxonomy)

table_by_taxa_age(epd.1, "Cedrus", 
                  as.character(seq(0, 21000, by=1000)))
}
}
