% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr_export_functions.R
\name{table_by_taxaAge}
\alias{table_by_taxaAge}
\title{Title TBW}
\usage{
table_by_taxaAge(agedcount, sample_label, taxa)
}
\arguments{
\item{agedcount}{TBW}

\item{sample_label}{TBW}

\item{taxa}{TBW}
}
\value{
TBW
}
\description{
Title TBW
}
\examples{
# TBW
}

