% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr-tableby_functions.R
\name{tableByTaxaAge}
\alias{tableByTaxaAge}
\title{Title TBW}
\usage{
tableByTaxaAge(agedcount, sample_label, taxa)
}
\arguments{
\item{agedcount}{TBW}

\item{sample_label}{TBW}

\item{taxa}{TBW}
}
\value{
TBW
}
\description{
Title TBW
}
\examples{
# TBW
}

