% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr-standardize_functions.R
\name{taxa2HigherTaxa}
\alias{taxa2HigherTaxa}
\title{Title TBW}
\usage{
taxa2HigherTaxa(count, epd_taxonomy)
}
\arguments{
\item{count}{TBW}

\item{epd_taxonomy}{TBW}
}
\value{
TBW
}
\description{
Title TBW
}
\examples{
# TBW
}

