% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr-standardize_methods.R
\docType{methods}
\name{taxa_to_highertaxa}
\alias{taxa_to_highertaxa}
\alias{taxa_to_highertaxa,epd.entity.df,data.frame-method}
\alias{taxa_to_highertaxa-method}
\alias{taxa_to_highertaxa,epd.entity,data.frame-method}
\alias{taxa_to_highertaxa-method}
\title{Change taxa to higher taxa level}
\usage{
taxa_to_highertaxa(x, epd.taxonomy)

\S4method{taxa_to_highertaxa}{epd.entity.df,data.frame}(x, epd.taxonomy)

\S4method{taxa_to_highertaxa}{epd.entity,data.frame}(x, epd.taxonomy)
}
\arguments{
\item{x}{\code{\link[EPDr:epd.entity.df]{epd.entity.df}} object.}

\item{epd.taxonomy}{Data frame with the taxonomy from the EPD as from the
\code{\link[EPDr:get_taxonomy_epd]{get_taxonomy_epd}} function.}
}
\value{
\code{\link[EPDr:epd.entity.df]{epd.entity.df}} object with new taxa names.
}
\description{
This function modifies the taxa names in the \code{@commdf@counts} slot of 
\code{\link[EPDr:epd.entity.df]{epd.entity.df}} objects.
More specifically this function compares the taxa name with the taxonomy of the EPD
to use the higher taxa names. If these changes result in duplicated columns of the same
taxa their values are unified by summing them.
}
\examples{
# epd.connection <- connectToEPD(host="localhost", database="epd",
# user="epdr", password="epdrpw")
# agedcounts.1 <- getAgedCounts(1, epd.connection)
# agedcounts.1@counts@taxa_names
# colnames(agedcounts.1@counts@counts)
# agedcounts.1.hn <- taxa_to_highertaxa(agedcounts.1, get_taxonomy_epd(epd.connection))
# agedcounts.1.hn@counts@taxa_names
# colnames(agedcounts.1.hn@counts@counts)
}
