% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr-standardize_methods.R
\docType{methods}
\name{taxa_to_highertaxa}
\alias{taxa_to_highertaxa}
\alias{taxa_to_highertaxa,epd.entity.df,data.frame-method}
\alias{taxa_to_highertaxa,epd.entity,data.frame-method}
\title{Change taxa to higher taxa level}
\usage{
taxa_to_highertaxa(x, epd.taxonomy, rm_null_mhvar = FALSE)

\S4method{taxa_to_highertaxa}{epd.entity.df,data.frame}(x, epd.taxonomy,
  rm_null_mhvar = FALSE)

\S4method{taxa_to_highertaxa}{epd.entity,data.frame}(x, epd.taxonomy,
  rm_null_mhvar = FALSE)
}
\arguments{
\item{x}{epd.entity \code{\link[EPDr]{epd.entity.df-class}} object.}

\item{epd.taxonomy}{data.frame Data frame with the taxonomy from 
the EPD as from the \code{\link[EPDr]{get_taxonomy_epd}} function.}

\item{rm_null_mhvar}{logical Logical value indicating whether to remove
or not those taxa that has no value for higher level taxa.}
}
\value{
Object of class \code{\link[EPDr]{epd.entity.df-class}} with new 
taxa names.
}
\description{
This function modifies the taxa names in the \code{@commdf@counts} slot
of \code{\link[EPDr]{epd.entity.df-class}} objects. More specifically, this 
function compares the taxa name with the taxonomy of the EPD to use 
the higher taxa names. If these changes result in duplicated columns of 
the same taxa their values are unified by summing them.
}
\examples{
\dontrun{
epd.connection <- connect_to_epd(host="localhost", database="epd",
                                 user="epdr", password="epdrpw")
epd.1 <- get_entity(1, epd.connection)
epd.1 <- entity_to_matrices(epd.1)
epd.1@commdf@taxanames
colnames(epd.1@commdf@counts)
epd.1.hn <- taxa_to_highertaxa(epd.1, get_taxonomy_epd(epd.connection))
epd.1.hn@commdf@taxanames
colnames(epd.1.hn@commdf@counts)
}
}
