% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr-manipulate_lists_functions.R
\name{unify_taxonomy}
\alias{unify_taxonomy}
\title{Unify taxonomy of counts in multiple objects}
\usage{
unify_taxonomy(list, epd.taxonomy, na_value = 0)
}
\arguments{
\item{list}{List of \code{\link[EPDr]{epd.entity.df-class}} objects to be modified.}

\item{epd.taxonomy}{Data frame with the taxonomy from the EPD as from the
\code{\link[EPDr]{get_taxonomy_epd}} function.}

\item{na_value}{numeric Number indicating the value to be used for
taxa not previously present in the entities.}
}
\value{
list of \code{\link[EPDr]{epd.entity.df-class}} objects with taxonomy
modified.
}
\description{
This function compares the taxa registered in a list of
\code{\link[EPDr]{epd.entity.df-class}} objects and expand their taxonomy (using
\code{\link[EPDr]{filter_taxa}}) to make them match. When neccesary the
function add empty columns (with \code{NA} values).
}
\examples{
\dontrun{
epd.connection <- connect_to_epd(host="localhost", database="epd",
                               user="epdr", password="epdrpw")
e.list <- listE(epd.connection, country=c("Spain"))
e.list <- e.list$e_
spain <- lapply(e.list, get_entity, epd.connection)
spain.ut <- unify_taxonomy(spain, get_taxonomy_epd(epd.connection))
colnames(spain[[1]]@commdf@counts)
colnames(spain.ut[[1]]@commdf@counts)
}
}
