% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr-standardize_functions.R
\name{unifyTaxonomy}
\alias{unifyTaxonomy}
\title{Title TBW}
\usage{
unifyTaxonomy(counts, epd_taxonomy)
}
\arguments{
\item{counts}{TBW}

\item{epd_taxonomy}{TBW}
}
\value{
TBW
}
\description{
Title TBW
}
\examples{
# TBW
}

