% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr.R
\name{c14toCLAM}
\alias{c14toCLAM}
\title{Reshape C14 data to CLAM format}
\usage{
c14toCLAM(C14)
}
\arguments{
\item{C14}{Data frame with C14 data as those extracted from \code{\link[EPDr:extractC14]{extractC14}}.}
}
\value{
Data frame with C14 data in CLAM format.
}
\description{
This function takes C14 data, as those extracted \code{\link[EPDr:extractC14]{extractC14}}, to fit into a new table that comply with
CLAM format.
}
\examples{
library(EPDr)
epd.connection <- connectToEPD(host="diegonl.ugr.es", database="epd_ddbb",
                               user="epdr", password="epdrpw")
c14 <- extractC14(400, epd.connection)
c14.clam <- c14toCLAM(c14)
}

