% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clam.R
\name{calibrate}
\alias{calibrate}
\title{Title #TBW}
\usage{
calibrate(cage = 2450, error = 50, reservoir = 0, prob = 0.95, cc = 1,
  cc1 = "IntCal13.14C", cc2 = "Marine13.14C", cc3 = "SHCal13.14C",
  cc4 = "mixed.14C", cc5 = "gluedHemispheres.14C", postbomb = FALSE,
  pb1 = "postbomb_NH1.14C", pb2 = "postbomb_NH2.14C",
  pb3 = "postbomb_NH3.14C", pb4 = "postbomb_SH1_2.14C",
  pb5 = "postbomb_SH3.14C", yrsteps = 1, pbsteps = 0.01, hpdsteps = 1,
  calibt = FALSE, yrmin = c(), yrmax = c(), minC14 = c(),
  maxC14 = c(), times = 5, calheight = 0.3, expand = 0.1,
  threshold = 1e-06, storedat = FALSE, graph = TRUE, xlab = c(),
  ylab = c(), BCAD = FALSE, mar = c(3.5, 3, 2, 1), mgp = c(2, 1, 0),
  bty = "l", title = c(), date.col = "red", cc.col = rgb(0, 0.5, 0,
  0.7), dist.col = rgb(0, 0, 0, 0.3), sd.col = rgb(0, 0, 0, 0.5))
}
\arguments{
\item{cage}{#TBW}

\item{error}{#TBW}

\item{reservoir}{#TBW}

\item{prob}{#TBW}

\item{cc}{#TBW}

\item{cc1}{#TBW}

\item{cc2}{#TBW}

\item{cc3}{#TBW}

\item{cc4}{#TBW}

\item{cc5}{#TBW}

\item{postbomb}{#TBW}

\item{pb1}{#TBW}

\item{pb2}{#TBW}

\item{pb3}{#TBW}

\item{pb4}{#TBW}

\item{pb5}{#TBW}

\item{yrsteps}{#TBW}

\item{pbsteps}{#TBW}

\item{hpdsteps}{#TBW}

\item{calibt}{#TBW}

\item{yrmin}{#TBW}

\item{yrmax}{#TBW}

\item{minC14}{#TBW}

\item{maxC14}{#TBW}

\item{times}{#TBW}

\item{calheight}{#TBW}

\item{expand}{#TBW}

\item{threshold}{#TBW}

\item{storedat}{#TBW}

\item{graph}{#TBW}

\item{xlab}{#TBW}

\item{ylab}{#TBW}

\item{BCAD}{#TBW}

\item{mar}{#TBW}

\item{mgp}{#TBW}

\item{bty}{#TBW}

\item{title}{#TBW}

\item{date.col}{#TBW}

\item{cc.col}{#TBW}

\item{dist.col}{#TBW}

\item{sd.col}{#TBW}
}
\value{
#TBW
}
\description{
Title #TBW
}
\examples{
 #TBW
}

