% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr-standardize_functions.R
\name{checkDefaultChronology}
\alias{checkDefaultChronology}
\title{Check default chronology on EPDr objects}
\usage{
checkDefaultChronology(object)
}
\arguments{
\item{object}{Any EPDr object but \code{\link[EPDr:counts]{counts}}.}
}
\value{
Logical value indicating whether the object has a default chronology (TRUE)
or not (FALSE).
}
\description{
The function check EPDr objects with slot @default_chronology 
(\code{\link[EPDr:ages]{ages}}, \code{\link[EPDr:agedcounts]{agedcounts}},
\code{\link[EPDr:chronology]{chronology}}, and \code{\link[EPDr:datation]{datation}})
to see if there are a default chronology specified.
}
\examples{
# epd.connection <- connectToEPD(host="localhost", database="epd",
#                                user="epdr", password="epdrpw")
# e.list <- listE(epd.connection, country=c("Spain"))
# e.list <- e.list$e_
# spain.agedcounts <- lapply(e.list, getAgedCounts, epd.connection)
# spain.agedcounts.wo <- sapply(spain.agedcounts, checkDefaultChronology)
# spain.agedcounts.wo
# spain.agedcounts[[1]]@ages@default_chronology
# spain.agedcounts[[16]]@ages@default_chronology

}
