% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr-standardize_functions.R
\name{checkRestriction}
\alias{checkRestriction}
\title{Check restrictions on EPDr objects}
\usage{
checkRestriction(object)
}
\arguments{
\item{object}{Any EPDr object (\code{\link[EPDr:ages]{ages}},
\code{\link[EPDr:counts]{counts}}, \code{\link[EPDr:agedcounts]{agedcounts}}, 
\code{\link[EPDr:chronology]{chronology}}, and \code{\link[EPDr:datation]{datation}}).}
}
\value{
logical value indicating whether the data are restricted (TRUE) or
unrestricted (FALSE).
}
\description{
This function check EPDr objects (\code{\link[EPDr:ages]{ages}},
\code{\link[EPDr:counts]{counts}}, \code{\link[EPDr:agedcounts]{agedcounts}}, 
\code{\link[EPDr:chronology]{chronology}}, and \code{\link[EPDr:datation]{datation}}), 
and return a logical value indicating whether the data are restricted (TRUE) or
unrestricted (FALSE).
}
\examples{
# epd.connection <- connectToEPD(host="localhost", database="epd",
# user="epdr", password="epdrpw")
# agedcount <- getAgedCounts(1, epd.connection)
# checkRestriction(agedcount)
# agedcount <- getAgedCounts(1046, epd.connection)
# checkRestriction(agedcount)

}
