% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr-classes.R
\docType{class}
\name{chronology-class}
\alias{chronology}
\alias{chronology-class}
\title{Title TBW}
\value{
TBW
}
\description{
Title TBW
}
\section{Slots}{

\describe{
\item{\code{e_}}{numeric. TBW}

\item{\code{restriction}}{data.frame. TBW}

\item{\code{entity}}{TBW}

\item{\code{site}}{data.frame. TBW}

\item{\code{number_of_chronologies}}{numeric. TBW}

\item{\code{default_chronology}}{numeric. TBW}

\item{\code{chron}}{data.frame. TBW}

\item{\code{agebasis}}{data.frame. TBW}
}}
\examples{
# TBW
}

