% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr.R
\name{core4Clam}
\alias{core4Clam}
\title{Title}
\usage{
core4Clam(core_number, connection, get_dephts = TRUE)
}
\arguments{
\item{core_number}{TBW}

\item{connection}{TBW}

\item{get_dephts}{TBW}
}
\value{
TBW
}
\description{
Title
}
\examples{
# TBW
}

