% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr-classes.R
\docType{class}
\name{counts-class}
\alias{counts}
\alias{counts-class}
\title{Class for Counts of an entity}
\description{
Class "counts" store in an organized and systematic way information about 
the particles counts in an specified entity in the European Pollen Database (EPD). This
object is created by \code{\link[EPDr:getCounts]{getCounts}}.
}
\details{
It has different elements all of which correspond to a unique entity
}
\section{Slots}{

\describe{
\item{\code{e_}}{numeric. The entity number (e_) as in the EPD.}

\item{\code{restriction}}{data.frame. Restriction of use information for that particular entity
in the database. It is important to know if we can freely use the data or should ask
for authorization}

\item{\code{counts_type}}{factor. Indicating the type of counts stored in slot \code{@counts}.
They can be one of two values "Counts" or "Percentages". "Counts" indicates that values
in \code{@counts} are raw counts as in the EPD. "Percentages" indicates that values in 
\code{@counts} are percentages calculated, for instance, with \code{\link[EPDr:trans2Percentages]{trans2Percentages}}.}

\item{\code{counts_processing}}{factor. Indicating whether the data in slot \code{@counts} has
been processed. It can be one of three values: "Samples", "Interpolated" or "Ranged 
means". "Samples" indicate that values in \code{@counts} correspond with the counts
for the palynological samples. "Interpolated" indicates that values in \code{@counts}
correspond to interpolated data at specific depth/ages among the palynological samples, 
using \code{\link[EPDr:interpolateCounts]{interpolateCounts}}. "Ranged means"
indicates that values in \code{@counts} represent mean values among all palynological samples for specified age/depth ranges calculated using \code{\link[EPDr:intervalsCounts]{intervalsCounts}}}

\item{\code{taxa_type}}{factor. Indicating the taxa used to calculate values in slot
\code{@counts}. The EPD allow establish three levels of taxonomy: the taxonomic level
determined by analyst when processing the entity, "accepted" name to resolve
synonymies, and "higher" to collapse taxa into higher taxonomical levels (i.e.,
species in the same genus, genus in the same family, etc). \code{taxa_type} can take 
any of three values: "Samples", "Accepted" or "Higher". "Samples" indicates that 
taxonomy in \code{@coutns} is the same the analyst submitted to the EPD. "Accepted"
indicates that taxonomy in \code{@counts} is modified using the accepted taxa according
to last review of the EPD. "Higher" indicates that taxonomy was modified to group
taxa into higher taxonomical levels. "Samples" is always found when data comes
directly from the EPD with \code{\link[EPDr:getCounts]{getCounts}}, whereas "Accepted" and "Higher" are specified when data are transformed using \code{\link[EPDr:taxa2AcceptedTaxa]{taxa2AcceptedTaxa}} and \code{\link[EPDr:taxa2HigherTaxa]{taxa2HigherTaxa}}
respectively.}

\item{\code{taxa_processing}}{factor. Indicating if taxonomy in \code{@counts} has been
processed. It can take any of three values: "Original", "Expanded", or "Taxized". 
"Original" indicates that taxonomy has not been modified. "Expanded" indicates that
taxonomy has been expanded beyond the taxa specified by the data provider, using
\code{\link[EPDr:filterTaxa]{filterTaxa}} or
\code{\link[EPDr:unifyTaxonomy]{unifyTaxonomy}}. "Taxized" indicates that taxonomy in
\code{@counts} has been resolved using the package \code{link[taxize]{taxize}}.
Function for this are on schedule but not implemented yet.}

\item{\code{entity}}{data.frame. Details of the entity.}

\item{\code{site}}{data.frame. Details of the site to which the entity belong to. Note that 
several entities can belong to the same site.}

\item{\code{taxa_names}}{character. Character vector with all taxa used in \code{@counts}.}

\item{\code{taxa_}}{numeric. Numeric vector indicating the taxa number for each of the taxa 
used in \code{@counts}.}

\item{\code{taxa_accepted}}{numeric. Numeric vector indicating the taxa number of 
corresponding accepted taxa for each taxa used in \code{@counts}.}

\item{\code{taxa_mhvar}}{numeric. Numeric vector indicating the taxa number of 
corresponding higher taxa for each taxa used in \code{@counts}.}

\item{\code{taxa_groupid}}{character. Character vector indicating the corresponding groupid
to which belong each taxa used in \code{@counts}.}

\item{\code{sample_}}{numeric. Numeric vector indicating the sample number of each sample used
in \code{@counts}. Sample here is each of the palynological samples in the entity.}

\item{\code{sample_label}}{character. Character vector indicating the sample name (or code)
of each sample used in \code{@counts}. Sample here is each of the palynological samples in the entity.}

\item{\code{default_ages}}{numeric. Numeric vector indicating the ages estimated for each
palynological sample according to the default chronology in the EPD for that entity.}

\item{\code{depthcm}}{numeric. Numeric vector indicating the depth (in cm) in which each
palynological sample was collected.}

\item{\code{counts}}{data.frame. Data frame with counts (raw counts or percentages) for each
taxon at each sample (original samples, interpolated or ranged). This takes the form
of a age/depth (rows) by taxon (columns).}
}}

