% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr-classes.R
\docType{class}
\name{datation-class}
\alias{datation-class}
\alias{datation}
\title{Class for Datation of an entity}
\description{
Class "datation" store all information about datation of an specified entity in the 
European Pollen Database (EPD). These objects are created by \code{\link[EPDr:getDatation]{getDatation}}.
}
\details{
Now, it include only C14 data and events, it should include in the future other
sort of data that are included in the EPD.

It has different elements all of which correspond to a unique entity in the database.
}
\section{Slots}{

\describe{
\item{\code{e_}}{numeric. The entity number (e_) as in the EPD.}

\item{\code{restriction}}{data.frame. Restriction of use information for that particular entity
in the database. It is important to know if we can freely use the data or should ask
for authorization}

\item{\code{entity}}{data.frame. Details of the entity.}

\item{\code{site}}{data.frame. Details of the site to which the entity belong to. Note that 
several entities can belong to the same site.}

\item{\code{postbomb_zone}}{factor. Indicate the postbomb zone in which the entity (site)
was sampled. This information is useful when calibrating radiocarbon dates with
\code{clam} or \code{bacon}.}

\item{\code{chronology}}{chronology. Object of class \code{\link{chronology}} for the entity.}

\item{\code{c14}}{data.frame. Details on radiocarbon (C14) data and analysis.}

\item{\code{events}}{data.frame. Details about events (e.g. tephra) that could appear in
the entity.}

\item{\code{depths}}{data.frame. Details about the samples (including depths) at which samples
were taken in the entity for palynological (pollen, spores, etc) analysis.}
}}

