% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDPr-DDBB_connection.R
\name{disconnectFromEPD}
\alias{disconnectFromEPD}
\title{Disconnect a connection to a EPD database
\code{disconnectFromEPD} turns down a connection to a EPD DDBB server.}
\usage{
disconnectFromEPD(connection = NULL)
}
\arguments{
\item{connection}{The connection object created with \code{\link[EPDr:connectToEPD]{connectToEPD}} to stablish the connection}
}
\value{
NULL It just disconnects from the EPD DDBB server and modifies the connection object to reflect the new status.
}
\description{
Disconnect a connection to a EPD database
\code{disconnectFromEPD} turns down a connection to a EPD DDBB server.
}
\examples{
#epd.connection <- connectToEPD(database="epd", user="epdr",
#                                 password="epdrpw", host="localhost")
#disconnectFromEPD(connection=epd.connection)
#epd.connection
}
