% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr-search_functions.R
\name{e_by_countries}
\alias{e_by_countries}
\title{Title TBW}
\usage{
e_by_countries(countries, connection, regions = NULL)
}
\arguments{
\item{countries}{TBW}

\item{connection}{TBW}

\item{regions}{TBW}
}
\value{
TBW
}
\description{
Title TBW
}
\examples{
# TBW
}

