% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr-search_functions.R
\name{e_by_worker}
\alias{e_by_worker}
\title{Title}
\usage{
e_by_worker(lastname, connection, firstname = NULL, initials = NULL)
}
\arguments{
\item{lastname}{}

\item{connection}{}

\item{initials}{}
}
\description{
Title
}

