% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr-standardize_functions.R
\name{filterTaxaGroups}
\alias{filterTaxaGroups}
\title{Title TBW}
\usage{
filterTaxaGroups(counts, taxa_groups)
}
\arguments{
\item{counts}{TBW}

\item{taxa_groups}{TBW}
}
\value{
TBW
}
\description{
Title TBW
}
\examples{
# TBW
}

