% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr-get_functions.R
\name{get_samples}
\alias{get_samples}
\alias{.get_psamples}
\alias{.get_pagedpt}
\alias{.get_pcounts}
\alias{.get_pvars}
\alias{.get_pvtrans}
\alias{.get_syntype}
\alias{.get_pgroup}
\alias{.get_groups}
\title{TBW}
\usage{
get_samples(e_, connection)

.get_psamples(e_, connection)

.get_pagedpt(e_, connection)

.get_pcounts(e_, connection)

.get_pvars(var_, connection)

.get_pvtrans(var_, connection)

.get_syntype(syntype, connection)

.get_pgroup(var_, connection)

.get_groups(groupid, connection)
}
\arguments{
\item{e_}{numeric. Value indicating the entity number (e_) of the database that want to be queried.}

\item{connection}{PostgreSQLConnection. Object of class \code{PostgreSQLConnection} as returned by function \code{\link[EPDr:connect_to_epd]{connect_to_epd}}.}

\item{var_}{TBW}

\item{syntype}{TBW}

\item{groupid}{TBW}
}
\value{
TBW
}
\description{
TBW
}
\section{get_samples}{

TBW
}

\section{.get_psamples}{

TBW
}

\section{.get_pagedpt}{

TBW
}

\section{.get_pcounts}{

TBW
}

\section{.get_pvars}{

TBW
}

\section{.get_pvtrans}{

TBW
}

\section{.get_syntype}{

TBW
}

\section{.get_pgroup}{

TBW
}

\section{.get_groups}{

TBW
}

\examples{
# Not run
# library(EPDr)
# epd.connection <- connect_to_epd(host = "localhost", database = "epd",
#                               user = "epdr", password = "epdrpw")
# chron.400 <- get_samples(400, epd.connection)
# chron.400
# disconnectFromEPD(epd.connection)


}
