% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr-get_functions.R
\name{getAges}
\alias{getAges}
\title{Query palynological-samples ages of EPD entities}
\usage{
getAges(e_, connection)
}
\arguments{
\item{e_}{numeric. Value indicating the entity number (e_) of the database that want to 
be queried.}

\item{connection}{PostgreSQLConnection. Object of class \code{PostgreSQLConnection} as
returned by function \code{\link[EPDr:connectToEPD]{connectToEPD}}.}
}
\value{
ages Object of class \code{\link[EPDr:ages]{ages}}.
}
\description{
This function queries the database to request information about estimated ages of palynological
samples of an specified entity. To perform the query the function requires the number
of the entity that want to be queried and a valid connection to the database. Hence,
the following parameters are mandatory:
}
\examples{
# Not run
# epd.connection <- connectToEPD(database="epd", user="epdr",
#                                 password="epdrpw", host="localhost")
# getAges(1, epd.connection)
# getAges(51, epd.connection)
# disconnectFromEPD(connection=epd.connection)

}
