% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr-get_functions.R
\name{getC14}
\alias{getC14}
\title{Extract C14 data for a particular core (entity in the EPD DDBB)}
\usage{
getC14(e_, connection)
}
\arguments{
\item{e_}{numeric. Value indicating the entity number (e_) of the database that want to 
be queried.}

\item{connection}{PostgreSQLConnection. Object of class \code{PostgreSQLConnection} as returned by function \code{\link[EPDr:connectToEPD]{connectToEPD}}.}
}
\value{
Data frame with combined information from c14 and geochron tables in the EPD.
}
\description{
Given a core number (as in the EPD DDBB: e_) the function returns a matrix with the C14 data associated to this core. This values
come from two different tables of the EPD: c14 and geochron.
}
\examples{
#epd.connection <- connectToEPD(database="epd", user="epdr",
#                                 password="epdrpw", host="localhost")
#getC14(1, epd.connection)
#getC14(400, epd.connection)
#disconnectFromEPD(connection=epd.connection)
}

