% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr-get_functions.R
\name{getChronology}
\alias{getChronology}
\title{Query chronologies of EPD entities}
\usage{
getChronology(e_, connection)
}
\arguments{
\item{e_}{numeric. Value indicating the entity number (e_) of the database that want to 
be queried.}

\item{connection}{PostgreSQLConnection. Object of class \code{PostgreSQLConnection} as
returned by function \code{\link[EPDr:connectToEPD]{connectToEPD}}.}
}
\value{
chronology. Object of class \code{\link[EPDr:chronology]{chronology}}. This object
store in an organized and systematic way all combined information from CHRON and AGEBASIS tables
in the EPD  (see documentation of the EPD:
\url{http://www.europeanpollendatabase.net/data/downloads/image/pollen-database-manual-20071011.doc}).
}
\description{
This function queries the database to request all information about the chronologies  
of an specified entity. A particular entity might have several chronologies developed in different
projects by differen researchers. To perform the query the function requires the number
of the entity that want to be queried and a valid connection to the database. Hence,
the following parameters are mandatory:
}
\examples{
# Not run
# epd.connection <- connectToEPD(database="epd", user="epdr",
#                                 password="epdrpw", host="localhost")
# getChronology(1, epd.connection)
# getChronology(400, epd.connection)
# disconnectFromEPD(connection=epd.connection)

}

