% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr-get_functions.R
\name{getDepths}
\alias{getDepths}
\title{Depths of pollen samples}
\usage{
getDepths(core_number, connection)
}
\arguments{
\item{core_number}{Integer with the core (entity) number for which C14 data want to be extracted.}

\item{connection}{Connection object to a EPD DDBB where the query is made.}
}
\value{
Data frame with all the information for pollen samples as in the p_sample table of the EPD DDBB.
}
\description{
Given a specific core (entity) this function return the information at which depths samples were taken for pollen data.
}
\examples{
#epd.connection <- connectToEPD(database="epd_ddbb", user="epdr",
#                                 password="epdrpw", host="diegonl.ugr.es")
#getDepths(1, epd.connection)
#getDepths(51, epd.connection)
#disconnectFromEPD(connection=epd.connection)
}

