% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr-get_functions.R
\name{getEntity}
\alias{getEntity}
\title{Query details of EPD Entities}
\usage{
getEntity(e_, connection)
}
\arguments{
\item{e_}{numeric. Value indicating the entity number (e_) of the database that want to 
be queried.}

\item{connection}{PostgreSQLConnection. Object of class \code{PostgreSQLConnection} as returned by function \code{\link[EPDr:connectToEPD]{connectToEPD}}.}
}
\value{
data.frame The function return a data.frame with all the information on the ENTITY table of the database (see documentation of the EPD: \url{http://www.europeanpollendatabase.net/data/downloads/image/pollen-database-manual-20071011.doc}).
}
\description{
This function query the database to return details of the specified entity. It requires 
the number of the entity that want to be queried and a valid connection to the database 
server, as returned by \code{\link[EPDr:connectToEPD]{connectToEPD}}. Hence, the 
following parameters are mandatory.
}
\examples{
# Not run
# library(EPDr)
# epd.connection <- connectToEPD(host="localhost", database="epd_ddbb",
#                               user="epdr", password="epdrpw")
# e.400 <- getEntity(400, epd.connection)
# e.400
# disconnectFromEPD(epd.connection)

}

