% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr-get_functions.R
\name{getSite}
\alias{getSite}
\title{Title TBW}
\usage{
getSite(core_number, connection)
}
\arguments{
\item{core_number}{TBW}

\item{connection}{TBW}
}
\value{
TBW
}
\description{
TBW
}
\examples{
#TBW
}

