% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clam.R
\name{glue.curves}
\alias{glue.curves}
\title{Title}
\usage{
glue.curves(nh = "IntCal09.14C", sh = "SHCal04.14C", offset = c(56, 24))
}
\arguments{
\item{nh}{}

\item{sh}{}

\item{offset}{}
}
\value{

}
\description{
Title
}

