% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr-standardize_functions.R
\name{intervalsCounts}
\alias{intervalsCounts}
\title{Title}
\usage{
intervalsCounts(agedcount, tmin, tmax, labels = NULL, chronology = NULL)
}
\arguments{
\item{agedcount}{}

\item{tmin}{}

\item{tmax}{}

\item{chronology}{}
}
\description{
Title
}

