% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr-list_functions.R
\name{list_e}
\alias{list_e}
\title{List entities in the EPD}
\usage{
list_e(connection, site = NULL, coords = NULL, lastname = NULL,
  firstname = NULL, initials = NULL, publ = NULL, country = NULL,
  region = NULL, restrictions = NULL, logical_operator = "AND")
}
\arguments{
\item{connection}{PostgreSQLConnection A valid connection to the 
database. See \code{\link[EPDr]{connect_to_epd}}.}

\item{site}{numeric Numeric or character vector indicating the site 
identification number or the site name to which entities will be retrieved.}

\item{coords}{numeric Numeric vector with four elements indicating the 
geographical coordinates to which the search is limited.}

\item{lastname}{character Character vector indicating the last name 
of data contributor for which entities will be retrieved.}

\item{firstname}{character Character vector indicating the first name 
of data contributor for which entities will be retrieved.}

\item{initials}{character Character vector indicating the initials 
of data contributor for which entities will be retrieved.}

\item{publ}{numeric Numeric vector indicating the publication 
identification number to which search is constrained.}

\item{country}{character Character vector indicating the three 
letters code or the full english name for the countries to which 
entities are desired.}

\item{region}{character Character vector indicating the two letters 
code or the full name for the regions to which entities are desired. 
To specify a region you have to specify one, and only one, country. 
You can specify multiple countries but the results  can be funky.}

\item{restrictions}{character Character letter ("U" or "R") specifying 
the usestatus of the data for the entities that are searched. U is 
for unrestricted and R is for restricted.}

\item{logical_operator}{character Character string indicating whether 
the effects of search criteria are additive ("OR") or substracting 
("AND"). If "OR" is specified, entities complying with any of the 
search criteria will be returned. If "AND" is specified, only entities 
that comply with all search criteria will be returned.}
}
\value{
This function return a data.frame with columns as in 
the \code{entity} table of the EPD database (see documentation of the EPD:
\url{http://www.europeanpollendatabase.net/data/downloads/image/pollen-database-manual-20071011.doc}). If search criteria are all empty, the function returns 
the whole table. If criteria are specified the table is complemented 
with columns from diferent tables in the database that store the 
required information to filter entities accordingly.
}
\description{
This function allows to search entities records by different search 
criteria that can be combined as desired.
}
\examples{
\dontrun{
epd.connection <- connect_to_epd(host = "localhost", database = "epd",
                              user = "epdr", password = "epdrpw")
list_e(epd.connection)
list_e(epd.connection, site = 1)
list_e(epd.connection, site = "Adange")
list_e(epd.connection, site = c(1:10))
list_e(epd.connection, country = "ESP")
list_e(epd.connection, country = "Spain")
list_e(epd.connection, country = "Spain", region = "Andalucia")
list_e(epd.connection, lastname = "Tzedakis")
list_e(epd.connection, restrictions = "R")
}
}
