% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr-list_functions.R
\name{list_publ}
\alias{list_publ}
\title{Title}
\usage{
list_publ(connection, e = NULL, countries = NULL, regions = NULL,
  sites = NULL, coords = NULL)
}
\arguments{
\item{connection}{}

\item{e}{}

\item{countries}{}

\item{regions}{}

\item{sites}{}

\item{coords}{}
}
\description{
Title
}

