% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr-list_functions.R
\name{list_regions}
\alias{list_regions}
\title{Title TBW}
\usage{
list_regions(connection, countries = NULL)
}
\arguments{
\item{connection}{TBW}

\item{countries}{TBW}
}
\value{
TBW
}
\description{
Title TBW
}
\examples{
# TBW
}

