% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr-list_functions.R
\name{list_sites}
\alias{list_sites}
\title{Title}
\usage{
list_sites(connection, countries = NULL, regions = NULL, coords = NULL)
}
\arguments{
\item{connection}{}

\item{countries}{}

\item{regions}{}

\item{coords}{}
}
\description{
Title
}

