% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr-list_functions.R
\name{list_taxagroups}
\alias{list_taxagroups}
\title{List taxa groups in the EPD}
\usage{
list_taxagroups(connection)
}
\arguments{
\item{connection}{PostgreSQLConnection. Object of class 
\code{PostgreSQLConnection} as returned by function 
\code{\link[EPDr]{connect_to_epd}}.}
}
\value{
data.frame The function returns a data frame with three columns: 
\code{groupid}, \code{groupcode}, and \code{groupname}.
}
\description{
This function looks into the database and returns all information in 
the GROUPS table (see documentation of the EPD:
\url{http://www.europeanpollendatabase.net/data/downloads/image/pollen-database-manual-20071011.doc}).
This is useful if the user want to restrict searches for particular 
groups of particles in the palynological samples. For instance, if somebody 
is only interested in Algae. Because the function return the whole table, 
it only requires a valid connection to the database server.
}
\examples{
# Not run
# library(EPDr)
# epd.connection <- connect_to_epd(host = "localhost", database = "epd",
#                                  user = "epdr", password = "epdrpw")
# list_taxagroups(epd.connection)

}
