% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr-list_functions.R
\name{listE}
\alias{listE}
\title{List entities in the EPD}
\usage{
listE(connection, site = NULL, coords = NULL, lastname = NULL,
  firstname = NULL, initials = NULL, publ = NULL, country = NULL,
  region = NULL, restrictions = NULL, logical_operator = "AND")
}
\arguments{
\item{connection}{A valid connection to the database see
\code{\link[EPDr:connectToEPD]{connectToEPD}}.}

\item{site}{Numeric or character vector indicating the site identification number or
the site name to which entities will be retrieved.}

\item{coords}{Numeric vector with four elements indicating the geographical coordinates
to which the search is limited.}

\item{lastname}{Character vector indicating the last name of data contributor for which
entities will be retrieved.}

\item{firstname}{Character vector indicating the first name of data contributor
for which entities will be retrieved.}

\item{initials}{Character vector indicating the initials of data contributor for which
entities will be retrieved.}

\item{publ}{Numeric vector indicating the publication identification number
to which search is constrained.}

\item{country}{Character vector indicating the three letters code or the full
english name for the countries to which entities are desired.}

\item{region}{Character vector indicating the two letters code or the full
name for the regions to which entities are desired. To specify a region you have to
specify one, and only one, country. You can specify multiple countries but the results 
can be funky.}

\item{restrictions}{Character letter ("U" or "R") specifying the usestatus of the data 
for the entities that are searched. U is for unrestricted and R is for restricted.}

\item{logical_operator}{Character string indicating whether the effects of 
search criteria are additive ("OR") or substracting ("AND"). If "OR" is specified,
entities complying with any of the search criteria will be returned. If "AND" is
specified, only entities that comply with all search criteria will be returned.}
}
\value{
This function return a data.frame with columns as in the \code{entity} table 
of the EPD database (see documentation of the EPD:
\url{http://www.europeanpollendatabase.net/data/downloads/image/pollen-database-manual-20071011.doc}). If search criteria are all empty, the function returns the whole
table. If criteria are specified the table is complemented with columns from diferent
tables in the database that store the required information to filter entities
accordingly.
}
\description{
This function allows to search entities records by different search criteria that can be
combined as desired.
}
\examples{
# Not run
# epd.connection <- connectToEPD(host="localhost", database="epd",
#                               user="epdr", password="epdrpw")
# listE(epd.connection)
# listE(epd.connection, site=1)
# listE(epd.connection, site="Adange")
# listE(epd.connection, site=c(1:10))
# listE(epd.connection, country="ESP")
# listE(epd.connection, country="Spain")
# listE(epd.connection, country="Spain", region="Andalucia")
# listE(epd.connection, lastname="Tzedakis")
# listE(epd.connection, restrictions="R")

}
