% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr-plotting_functions.R
\name{map_by_TaxaAge}
\alias{map_by_TaxaAge}
\title{Title}
\usage{
map_by_TaxaAge(agedcounts, taxa, sample_label, pres_abse = FALSE,
  pollen_thres = NULL, zoom_coord = NULL, points_pch = 21,
  points_colour = NULL, points_fill = NULL, points_range_size = NULL,
  map_title = NULL, legend_range = NULL, legend_title = NULL,
  napoints_size = 0.75, napoints_pch = 19, napoints_colour = "grey45",
  napoints_fill = NA, countries_bg_colour = "grey80",
  countries_border_colour = "grey90")
}
\arguments{
\item{agedcounts}{}

\item{taxa}{}

\item{sample_label}{}

\item{pres_abse}{}

\item{pollen_thres}{}

\item{zoom_coord}{}

\item{points_pch}{}

\item{points_colour}{}

\item{points_fill}{}

\item{points_range_size}{}

\item{map_title}{}

\item{legend_range}{}

\item{legend_title}{}

\item{napoints_size}{}

\item{napoints_pch}{}

\item{napoints_colour}{}

\item{napoints_fill}{}

\item{countries_bg_colour}{}

\item{countries_border_colour}{}
}
\description{
Title
}

