% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clam.R
\name{plot.proxies}
\alias{plot.proxies}
\title{Title}
\usage{
\method{plot}{proxies}(prox, errors = TRUE, proxcol = grey(0.5),
  revyr = TRUE)
}
\arguments{
\item{prox}{}

\item{errors}{}

\item{proxcol}{}

\item{revyr}{}
}
\description{
Title
}

