% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr-manipulate_lists_functions.R
\name{remove_restricted}
\alias{remove_restricted}
\title{Remove restricted data from a list of objects}
\usage{
remove_restricted(list)
}
\arguments{
\item{list}{List of EPDr objects (\code{\link[EPDr:epd.entity.df]{epd.entity.df}}, or
\code{\link[EPDr:epd.entity]{epd.entity}}).}
}
\value{
List of EPDr objects.
}
\description{
This function is designed to work with list of EPDr objects
(\code{\link[EPDr:epd.entity.df]{epd.entity.df}}). The function parses all the elements
on the list and remove those with restriction on their use.
}
\examples{
# epd.connection <- connect_to_epd(host="localhost", database="epd",
# user="epdr", password="epdrpw")
# e.list <- listE(epd.connection, country=c("Spain"))
# e.list <- e.list$e_
# spain.agedcounts <- lapply(e.list, getAgedCounts, epd.connection)
# spain.agedcounts.un <- remove_restricted(spain.agedcounts)
# length(spain.agedcounts)
# length(spain.agedcounts.un)

}
