% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr-manipulate_lists_functions.R
\name{remove_wo_ages}
\alias{remove_wo_ages}
\title{Remove data without ages from a list of objects}
\usage{
remove_wo_ages(list)
}
\arguments{
\item{list}{List of EPDr objects (\code{\link[EPDr]{epd.entity.df}},
or \code{\link[EPDr]{epd.entity}})}
}
\value{
List of EPDr objects.
}
\description{
This function is designed to work with list of EPDr objects
(\code{\link[EPDr]{epd.entity.df}}, or
\code{\link[EPDr]{epd.entity}}).
The function parses all the elements on the list and remove those without a default
chronology (see \code{\link[EPDr]{check_default_chron}}).
}
\examples{
# epd.connection <- connect_to_epd(host="localhost", database="epd",
#                                user="epdr", password="epdrpw")
# e.list <- listE(epd.connection, country=c("Spain"))
# e.list <- e.list$e_
# spain.agedcounts <- lapply(e.list, getAgedCounts, epd.connection)
# spain.agedcounts.wo <- remove_wo_ages(spain.agedcounts)
# length(spain.agedcounts)
# length(spain.agedcounts.wo)

}
