% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr-standardize_functions.R
\name{removeRestricted}
\alias{removeRestricted}
\title{Title TBW}
\usage{
removeRestricted(agedcounts)
}
\arguments{
\item{agedcounts}{TBW}
}
\value{
TBW
}
\description{
Title TBW
}
\examples{
# TBW
}

