% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr-standardize_functions.R
\name{removeWithoutAges}
\alias{removeWithoutAges}
\title{Title TBW}
\usage{
removeWithoutAges(agedcounts)
}
\arguments{
\item{agedcounts}{TBW}
}
\value{
TBW
}
\description{
Title TBW
}
\examples{
# TBW
}

