% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clam.R
\name{student.t}
\alias{student.t}
\title{Title}
\usage{
student.t(y = 2450, error = 50, t.a = 3, t.b = 4, cc = 1,
  postbomb = c(), cc1 = "IntCal13", cc2 = "Marine13", cc3 = "SHCal13",
  cc4 = "ConstCal", Cutoff = 1e-05)
}
\arguments{
\item{y}{}

\item{error}{}

\item{t.a}{}

\item{t.b}{}

\item{cc}{}

\item{postbomb}{}

\item{cc1}{}

\item{cc2}{}

\item{cc3}{}

\item{cc4}{}

\item{Cutoff}{}
}
\value{

}
\description{
Title
}

