% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr_export_functions.R
\name{table_by_taxaAge}
\alias{table_by_taxaAge}
\title{Title}
\usage{
table_by_taxaAge(agedcount, sample_label, taxa)
}
\arguments{
\item{agedcount}{}

\item{sample_label}{}

\item{taxa}{}
}
\description{
Title
}

