% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr-tableby_functions.R
\name{tableByTaxaAge}
\alias{tableByTaxaAge}
\title{Tabulate counts by taxa and age}
\usage{
tableByTaxaAge(agedcounts, sample_label, taxa)
}
\arguments{
\item{agedcounts}{\code{\link[EPDr:agedcounts]{agedcounts}} object where to extract the 
data from.}

\item{sample_label}{Character vector indicating the ages or time intervals to be included
in the table.}

\item{taxa}{Character vector indicating the taxa to be included in the table. Several
taxa can be specified but the function returns data for only one taxa by summing all
counts, and the taxa name specified in the output is the first one in \code{taxa}.}
}
\value{
Data frame with five columns:
\itemize{
\item{"e_"}{Entity identification number.}
\item{"londd"}{Longitude of the site in decimal degrees.}
\item{"latdd"}{Latitude of the site in decimal degrees.}
\item{"count"}{The count of that taxon in that particular sample (age or time interval).}
\item{"sample_label"}{The sample (age or time interval) at which the particulates
were counted.}
\item{"taxa_label"}{The taxa that has been counted.}
}
}
\description{
This function tabulate data from the \code{@counts} slot in EPDr objects
(\code{\link[EPDr:agedcounts]{agedcounts}} or \code{\link[EPDr:counts]{counts}} to
summarize information for particular taxa at particular age or time intervals (samples).
This function is useful to reshape the data to be plotted and mapped by \code{link[ggplot]{ggplot}}. It was written to be used by \code{\link[EPDr:mapTaxaAge]{mapTaxaAge}} function.
}
\examples{
# TBW
}
