% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPDr-standardize_functions.R
\name{trans2Percentages}
\alias{trans2Percentages}
\title{Title TBW}
\usage{
trans2Percentages(counts)
}
\arguments{
\item{counts}{TBW}
}
\value{
TBW
}
\description{
Title TBW
}
\examples{
# TBW
}

